Refactor coordinate generator to accept any borough
