--- server/src/yelp-services/constants/manhattan-areas.ts
@@ -1,4390 +1,0 @@
-type Feature = {
-  type: string;
-  properties: {
-    cartodb_id: number;
-    name: string;
-    created_at: string;
-    updated_at: string;
-  };
-  geometry: {
-    type: string;
-    coordinates: number[][][][];
-  };
-};
-
-type GeoJson = {
-  type: string;
-  features: Feature[];
-};
-
-// Source: https://github.com/codeforgermany/click_that_hood/blob/main/public/data/manhattan.geojson
-export const CLICK_THAT_HOOD_MANHATTAN: GeoJson = {
-  type: "FeatureCollection",
-  features: [
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 9,
-        name: "Battery Park City",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-74.013754, 40.71369],
-              [-74.014262, 40.710928],
-              [-74.016542, 40.704733],
-              [-74.016174, 40.702572],
-              [-74.01563, 40.701862],
-              [-74.015127, 40.70139],
-              [-74.014008, 40.701043],
-              [-74.013959, 40.700966],
-              [-74.014105, 40.700511],
-              [-74.014091, 40.700508],
-              [-74.014149, 40.700321],
-              [-74.014221, 40.700111],
-              [-74.01431, 40.700121],
-              [-74.01431, 40.700134],
-              [-74.014178, 40.700541],
-              [-74.014165, 40.700539],
-              [-74.014152, 40.700578],
-              [-74.014289, 40.700644],
-              [-74.014297, 40.700629],
-              [-74.014315, 40.700611],
-              [-74.01456, 40.70031],
-              [-74.014633, 40.700197],
-              [-74.014671, 40.70021],
-              [-74.014667, 40.700239],
-              [-74.014608, 40.700324],
-              [-74.014324, 40.700661],
-              [-74.014643, 40.700817],
-              [-74.015068, 40.700318],
-              [-74.015098, 40.700334],
-              [-74.014674, 40.700835],
-              [-74.014811, 40.700904],
-              [-74.014828, 40.700895],
-              [-74.015021, 40.700993],
-              [-74.015142, 40.700843],
-              [-74.015678, 40.701256],
-              [-74.016139, 40.701646],
-              [-74.016533, 40.702037],
-              [-74.016819, 40.702339],
-              [-74.017145, 40.702705],
-              [-74.017252, 40.702836],
-              [-74.017398, 40.703019],
-              [-74.01744, 40.703078],
-              [-74.01754, 40.703236],
-              [-74.017575, 40.703308],
-              [-74.017617, 40.703423],
-              [-74.017626, 40.703423],
-              [-74.017648, 40.703495],
-              [-74.017671, 40.703606],
-              [-74.017683, 40.703711],
-              [-74.017673, 40.703712],
-              [-74.017677, 40.703771],
-              [-74.017674, 40.703834],
-              [-74.017666, 40.703899],
-              [-74.017658, 40.703933],
-              [-74.017643, 40.703989],
-              [-74.017654, 40.703998],
-              [-74.017655, 40.70401],
-              [-74.017644, 40.704022],
-              [-74.017634, 40.704025],
-              [-74.017627, 40.704024],
-              [-74.01757, 40.704184],
-              [-74.017795, 40.704235],
-              [-74.017856, 40.704079],
-              [-74.017751, 40.704055],
-              [-74.017791, 40.703951],
-              [-74.017818, 40.703938],
-              [-74.017838, 40.703963],
-              [-74.017808, 40.704032],
-              [-74.017912, 40.704054],
-              [-74.017943, 40.703973],
-              [-74.017955, 40.703976],
-              [-74.017844, 40.704242],
-              [-74.018448, 40.704162],
-              [-74.018668, 40.704435],
-              [-74.018656, 40.704441],
-              [-74.017516, 40.704581],
-              [-74.017481, 40.704667],
-              [-74.018043, 40.704628],
-              [-74.018054, 40.70463],
-              [-74.018065, 40.704636],
-              [-74.018113, 40.704554],
-              [-74.018863, 40.70472],
-              [-74.018887, 40.704729],
-              [-74.018926, 40.704774],
-              [-74.019343, 40.706094],
-              [-74.019123, 40.706976],
-              [-74.019103, 40.707],
-              [-74.019066, 40.707014],
-              [-74.019033, 40.707013],
-              [-74.018954, 40.706997],
-              [-74.018946, 40.707018],
-              [-74.018937, 40.707034],
-              [-74.018909, 40.707064],
-              [-74.018885, 40.707086],
-              [-74.018859, 40.7071],
-              [-74.018829, 40.707115],
-              [-74.018796, 40.707128],
-              [-74.018761, 40.707138],
-              [-74.018726, 40.707143],
-              [-74.01869, 40.707142],
-              [-74.018655, 40.707142],
-              [-74.018631, 40.70714],
-              [-74.018581, 40.707124],
-              [-74.018547, 40.707108],
-              [-74.018571, 40.70708],
-              [-74.018607, 40.707095],
-              [-74.018644, 40.707105],
-              [-74.018648, 40.707082],
-              [-74.018687, 40.707084],
-              [-74.018721, 40.707082],
-              [-74.018751, 40.70708],
-              [-74.018762, 40.707076],
-              [-74.01875, 40.70705],
-              [-74.018774, 40.707044],
-              [-74.018794, 40.707032],
-              [-74.01882, 40.707016],
-              [-74.018848, 40.706983],
-              [-74.018849, 40.70697],
-              [-74.018811, 40.706966],
-              [-74.018799, 40.706982],
-              [-74.018776, 40.706997],
-              [-74.018742, 40.707008],
-              [-74.018679, 40.707019],
-              [-74.018648, 40.707018],
-              [-74.018628, 40.707011],
-              [-74.018605, 40.706989],
-              [-74.018572, 40.706938],
-              [-74.018565, 40.706906],
-              [-74.018549, 40.706908],
-              [-74.018535, 40.706918],
-              [-74.01853, 40.706951],
-              [-74.018217, 40.707742],
-              [-74.018217, 40.707751],
-              [-74.01819, 40.707822],
-              [-74.018656, 40.707924],
-              [-74.018647, 40.707947],
-              [-74.018801, 40.707981],
-              [-74.018832, 40.707999],
-              [-74.018853, 40.708022],
-              [-74.018855, 40.708048],
-              [-74.018608, 40.709135],
-              [-74.017873, 40.712315],
-              [-74.01777, 40.712835],
-              [-74.017753, 40.712848],
-              [-74.017732, 40.712849],
-              [-74.01772, 40.712845],
-              [-74.017721, 40.712784],
-              [-74.0178, 40.712345],
-              [-74.016625, 40.712157],
-              [-74.016545, 40.712456],
-              [-74.016592, 40.712548],
-              [-74.016619, 40.71265],
-              [-74.016618, 40.712756],
-              [-74.016591, 40.712862],
-              [-74.016555, 40.712929],
-              [-74.016541, 40.712956],
-              [-74.016477, 40.713029],
-              [-74.016392, 40.7131],
-              [-74.01632, 40.713408],
-              [-74.016541, 40.713441],
-              [-74.016911, 40.713496],
-              [-74.017544, 40.713611],
-              [-74.017568, 40.713524],
-              [-74.017579, 40.713443],
-              [-74.017648, 40.713206],
-              [-74.01767, 40.713069],
-              [-74.017673, 40.713066],
-              [-74.017676, 40.713063],
-              [-74.017679, 40.713061],
-              [-74.017683, 40.713059],
-              [-74.017688, 40.713057],
-              [-74.017692, 40.713056],
-              [-74.017697, 40.713056],
-              [-74.017702, 40.713056],
-              [-74.017706, 40.713057],
-              [-74.01771, 40.713058],
-              [-74.017714, 40.71306],
-              [-74.017718, 40.713062],
-              [-74.017721, 40.713064],
-              [-74.017723, 40.713067],
-              [-74.017725, 40.71307],
-              [-74.017657, 40.71344],
-              [-74.01671, 40.718624],
-              [-74.016694, 40.718649],
-              [-74.016659, 40.718666],
-              [-74.01663, 40.718665],
-              [-74.015947, 40.718576],
-              [-74.015944, 40.718589],
-              [-74.015669, 40.718554],
-              [-74.015671, 40.718538],
-              [-74.015001, 40.718449],
-              [-74.014996, 40.718473],
-              [-74.01499, 40.718507],
-              [-74.014983, 40.718524],
-              [-74.01497, 40.718536],
-              [-74.014945, 40.718544],
-              [-74.014703, 40.718516],
-              [-74.013213, 40.718316],
-              [-74.013072, 40.718939],
-              [-74.013036, 40.719086],
-              [-74.01244, 40.719058],
-              [-74.013754, 40.71369],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 50,
-        name: "Central Park",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.949657, 40.797069],
-              [-73.949641, 40.796804],
-              [-73.949552, 40.796693],
-              [-73.9494, 40.796638],
-              [-73.949695, 40.796217],
-              [-73.95576, 40.787907],
-              [-73.973015, 40.764279],
-              [-73.98115, 40.767754],
-              [-73.981451, 40.767983],
-              [-73.981468, 40.76815],
-              [-73.981565, 40.768307],
-              [-73.981549, 40.768559],
-              [-73.981439, 40.768763],
-              [-73.958358, 40.800369],
-              [-73.958074, 40.800359],
-              [-73.95785, 40.800528],
-              [-73.949657, 40.797069],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 52,
-        name: "Chelsea",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.993464, 40.75219],
-              [-73.987937, 40.749855],
-              [-73.987933, 40.749853],
-              [-73.987973, 40.749467],
-              [-73.991356, 40.744763],
-              [-73.994211, 40.741013],
-              [-73.989777, 40.739169],
-              [-73.990371, 40.737098],
-              [-73.991725, 40.735245],
-              [-73.996797, 40.737364],
-              [-74.00881, 40.742378],
-              [-74.009161, 40.742509],
-              [-74.009092, 40.742886],
-              [-74.00924, 40.7429],
-              [-74.00962, 40.742951],
-              [-74.00962, 40.742952],
-              [-74.009568, 40.743258],
-              [-74.009585, 40.743264],
-              [-74.012119, 40.743517],
-              [-74.012137, 40.743532],
-              [-74.012148, 40.743559],
-              [-74.01209, 40.743898],
-              [-74.012069, 40.743922],
-              [-74.012021, 40.74394],
-              [-74.009512, 40.743676],
-              [-74.009489, 40.743696],
-              [-74.009438, 40.743971],
-              [-74.009394, 40.743969],
-              [-74.0089, 40.743918],
-              [-74.00873, 40.744882],
-              [-74.009105, 40.744931],
-              [-74.009612, 40.744998],
-              [-74.009558, 40.745289],
-              [-74.009917, 40.745327],
-              [-74.009934, 40.745317],
-              [-74.009998, 40.744955],
-              [-74.009987, 40.744944],
-              [-74.009869, 40.744934],
-              [-74.009872, 40.744917],
-              [-74.010168, 40.744942],
-              [-74.010165, 40.744964],
-              [-74.010054, 40.744954],
-              [-74.010038, 40.744963],
-              [-74.009974, 40.745323],
-              [-74.009991, 40.745337],
-              [-74.010469, 40.745387],
-              [-74.010488, 40.745376],
-              [-74.01055, 40.745013],
-              [-74.010532, 40.745003],
-              [-74.010417, 40.744994],
-              [-74.010419, 40.744977],
-              [-74.010764, 40.745005],
-              [-74.010762, 40.745024],
-              [-74.010606, 40.745011],
-              [-74.010586, 40.74502],
-              [-74.010531, 40.745382],
-              [-74.010545, 40.745392],
-              [-74.011104, 40.745448],
-              [-74.011116, 40.745436],
-              [-74.011172, 40.745078],
-              [-74.011153, 40.745069],
-              [-74.011043, 40.745061],
-              [-74.011045, 40.745046],
-              [-74.011388, 40.745073],
-              [-74.011386, 40.745088],
-              [-74.01123, 40.745076],
-              [-74.011154, 40.745446],
-              [-74.011167, 40.74546],
-              [-74.01174, 40.74552],
-              [-74.011795, 40.745155],
-              [-74.011783, 40.745135],
-              [-74.011669, 40.745127],
-              [-74.011671, 40.745114],
-              [-74.011836, 40.745126],
-              [-74.011766, 40.745578],
-              [-74.011783, 40.745602],
-              [-74.011719, 40.74594],
-              [-74.011711, 40.745961],
-              [-74.011679, 40.745979],
-              [-74.011635, 40.745978],
-              [-74.010555, 40.745871],
-              [-74.010551, 40.745903],
-              [-74.010128, 40.745857],
-              [-74.010128, 40.745842],
-              [-74.009947, 40.745817],
-              [-74.009948, 40.7458],
-              [-74.009464, 40.74575],
-              [-74.009366, 40.746337],
-              [-74.011528, 40.746575],
-              [-74.011562, 40.746587],
-              [-74.011585, 40.746609],
-              [-74.011582, 40.746636],
-              [-74.011525, 40.746973],
-              [-74.011479, 40.746985],
-              [-74.011476, 40.746997],
-              [-74.009243, 40.746763],
-              [-74.009137, 40.74736],
-              [-74.00956, 40.747403],
-              [-74.00955, 40.747371],
-              [-74.009552, 40.747339],
-              [-74.009964, 40.747378],
-              [-74.009958, 40.747412],
-              [-74.009902, 40.747438],
-              [-74.011357, 40.747592],
-              [-74.011395, 40.747611],
-              [-74.011407, 40.747631],
-              [-74.011407, 40.747631],
-              [-74.011347, 40.747976],
-              [-74.011332, 40.748001],
-              [-74.011304, 40.748015],
-              [-74.011266, 40.748014],
-              [-74.009058, 40.747786],
-              [-74.008946, 40.748412],
-              [-74.008979, 40.748413],
-              [-74.008979, 40.7484],
-              [-74.009071, 40.74841],
-              [-74.009074, 40.748395],
-              [-74.009415, 40.74843],
-              [-74.009416, 40.748437],
-              [-74.009523, 40.748448],
-              [-74.009519, 40.748474],
-              [-74.011167, 40.748645],
-              [-74.011195, 40.748657],
-              [-74.011216, 40.748678],
-              [-74.011216, 40.748697],
-              [-74.011149, 40.749026],
-              [-74.011133, 40.749042],
-              [-74.011109, 40.74905],
-              [-74.011078, 40.749052],
-              [-74.009994, 40.748937],
-              [-74.009948, 40.748932],
-              [-74.00956, 40.749935],
-              [-74.009723, 40.749972],
-              [-74.009728, 40.749964],
-              [-74.010192, 40.750109],
-              [-74.010808, 40.750319],
-              [-74.010757, 40.750411],
-              [-74.009683, 40.750063],
-              [-74.009694, 40.750044],
-              [-74.009542, 40.750001],
-              [-74.009518, 40.750107],
-              [-74.009253, 40.749985],
-              [-74.009119, 40.750332],
-              [-74.010674, 40.751072],
-              [-74.010591, 40.751323],
-              [-74.009001, 40.750648],
-              [-74.008735, 40.751349],
-              [-74.008731, 40.751466],
-              [-74.008667, 40.751628],
-              [-74.008533, 40.751884],
-              [-74.00876, 40.751948],
-              [-74.009078, 40.752069],
-              [-74.010217, 40.752548],
-              [-74.010125, 40.752669],
-              [-74.008997, 40.752189],
-              [-74.008915, 40.752164],
-              [-74.008591, 40.752028],
-              [-74.008581, 40.752044],
-              [-74.009341, 40.752356],
-              [-74.009328, 40.752374],
-              [-74.008479, 40.752019],
-              [-74.008426, 40.752158],
-              [-74.009991, 40.752816],
-              [-74.00992, 40.752912],
-              [-74.009971, 40.752934],
-              [-74.009919, 40.753004],
-              [-74.009789, 40.752948],
-              [-74.009888, 40.752814],
-              [-74.009839, 40.752795],
-              [-74.009831, 40.752803],
-              [-74.00883, 40.75238],
-              [-74.008836, 40.752372],
-              [-74.008792, 40.752352],
-              [-74.008688, 40.752494],
-              [-74.008354, 40.752351],
-              [-74.007811, 40.753779],
-              [-74.007761, 40.75391],
-              [-74.00762, 40.754281],
-              [-74.007745, 40.754337],
-              [-74.007597, 40.75453],
-              [-74.007472, 40.754477],
-              [-74.007285, 40.754736],
-              [-74.007414, 40.754789],
-              [-74.007272, 40.754986],
-              [-74.007145, 40.754932],
-              [-74.006394, 40.755972],
-              [-74.005567, 40.757118],
-              [-74.004927, 40.757023],
-              [-73.993464, 40.75219],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 54,
-        name: "Chinatown",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.999991, 40.717977],
-              [-73.998636, 40.717077],
-              [-73.997791, 40.716793],
-              [-73.996478, 40.719048],
-              [-73.994808, 40.718457],
-              [-73.989155, 40.716708],
-              [-73.9902, 40.714665],
-              [-73.99012, 40.713382],
-              [-73.992623, 40.713168],
-              [-73.997994, 40.712722],
-              [-73.998174, 40.713328],
-              [-73.998086, 40.713426],
-              [-73.998535, 40.713446],
-              [-74.00045, 40.714365],
-              [-74.000528, 40.71517],
-              [-74.000789, 40.715291],
-              [-73.99994, 40.716553],
-              [-74.000796, 40.716907],
-              [-74.001554, 40.717305],
-              [-74.00208, 40.716731],
-              [-74.003523, 40.717444],
-              [-74.003533, 40.717449],
-              [-74.001886, 40.719395],
-              [-73.999991, 40.717977],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 57,
-        name: "Civic Center",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.997994, 40.712722],
-              [-73.99777, 40.711971],
-              [-74.002363, 40.711591],
-              [-74.003337, 40.711236],
-              [-74.003355, 40.71123],
-              [-74.003662, 40.711487],
-              [-74.005358, 40.712151],
-              [-74.005769, 40.712007],
-              [-74.008614, 40.711372],
-              [-74.003533, 40.717449],
-              [-74.003523, 40.717444],
-              [-74.00208, 40.716731],
-              [-74.001554, 40.717305],
-              [-74.000796, 40.716907],
-              [-73.99994, 40.716553],
-              [-74.000789, 40.715291],
-              [-74.000528, 40.71517],
-              [-74.00045, 40.714365],
-              [-73.998535, 40.713446],
-              [-73.998086, 40.713426],
-              [-73.998174, 40.713328],
-              [-73.997994, 40.712722],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 84,
-        name: "East Harlem",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.93383, 40.819521],
-              [-73.93383, 40.819521],
-              [-73.933829, 40.815585],
-              [-73.933945, 40.815588],
-              [-73.933945, 40.815588],
-              [-73.933946, 40.815588],
-              [-73.934015, 40.814484],
-              [-73.933857, 40.814481],
-              [-73.933865, 40.814153],
-              [-73.934339, 40.809562],
-              [-73.934121, 40.808656],
-              [-73.933561, 40.807601],
-              [-73.931159, 40.804967],
-              [-73.930309, 40.803786],
-              [-73.929208, 40.801148],
-              [-73.929035, 40.801081],
-              [-73.928902, 40.799855],
-              [-73.929036, 40.796763],
-              [-73.929279, 40.795853],
-              [-73.929875, 40.795035],
-              [-73.929842, 40.795015],
-              [-73.929851, 40.795001],
-              [-73.92981, 40.794983],
-              [-73.929833, 40.794949],
-              [-73.929887, 40.794967],
-              [-73.929901, 40.794947],
-              [-73.929916, 40.794949],
-              [-73.929937, 40.794919],
-              [-73.92992, 40.79491],
-              [-73.930069, 40.794701],
-              [-73.930108, 40.794713],
-              [-73.930197, 40.794593],
-              [-73.933485, 40.792774],
-              [-73.935265, 40.791526],
-              [-73.93519, 40.791471],
-              [-73.935468, 40.791264],
-              [-73.935539, 40.791319],
-              [-73.935831, 40.791095],
-              [-73.937062, 40.789414],
-              [-73.937053, 40.789406],
-              [-73.937081, 40.789366],
-              [-73.937072, 40.789346],
-              [-73.936141, 40.788949],
-              [-73.936128, 40.788939],
-              [-73.936127, 40.788927],
-              [-73.936219, 40.7888],
-              [-73.936228, 40.788796],
-              [-73.936243, 40.788796],
-              [-73.936883, 40.789059],
-              [-73.936946, 40.789085],
-              [-73.937148, 40.789168],
-              [-73.937148, 40.789168],
-              [-73.93715, 40.789168],
-              [-73.937774, 40.787783],
-              [-73.938116, 40.787185],
-              [-73.939826, 40.785257],
-              [-73.942435, 40.783918],
-              [-73.94324, 40.783328],
-              [-73.943592, 40.782748],
-              [-73.944023, 40.78296],
-              [-73.95576, 40.787907],
-              [-73.949695, 40.796217],
-              [-73.9494, 40.796638],
-              [-73.949552, 40.796693],
-              [-73.949641, 40.796804],
-              [-73.949657, 40.797069],
-              [-73.949374, 40.797158],
-              [-73.949228, 40.797149],
-              [-73.949073, 40.797094],
-              [-73.944597, 40.803228],
-              [-73.942993, 40.802561],
-              [-73.941166, 40.805063],
-              [-73.942754, 40.805744],
-              [-73.933992, 40.817786],
-              [-73.934289, 40.819569],
-              [-73.93383, 40.819521],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 87,
-        name: "East Village",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.990704, 40.734782],
-              [-73.982554, 40.731375],
-              [-73.982553, 40.731375],
-              [-73.971629, 40.726761],
-              [-73.97163, 40.72676],
-              [-73.971768, 40.725833],
-              [-73.971774, 40.725821],
-              [-73.971791, 40.725814],
-              [-73.971817, 40.725815],
-              [-73.972188, 40.724045],
-              [-73.972852, 40.721247],
-              [-73.973383, 40.719189],
-              [-73.973513, 40.718775],
-              [-73.973576, 40.718589],
-              [-73.974739, 40.718801],
-              [-73.975452, 40.718947],
-              [-73.977224, 40.71931],
-              [-73.978753, 40.719934],
-              [-73.992604, 40.724136],
-              [-73.991285, 40.727762],
-              [-73.990757, 40.729741],
-              [-73.992854, 40.730083],
-              [-73.99146, 40.731738],
-              [-73.990704, 40.734782],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 91,
-        name: "Ellis Island",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-74.041661, 40.696453],
-              [-74.043674, 40.69802],
-              [-74.042704, 40.698801],
-              [-74.042965, 40.698999],
-              [-74.041099, 40.700493],
-              [-74.040801, 40.700269],
-              [-74.040313, 40.700626],
-              [-74.040391, 40.700687],
-              [-74.040412, 40.700678],
-              [-74.040536, 40.700782],
-              [-74.040481, 40.700821],
-              [-74.04035, 40.700716],
-              [-74.040365, 40.700704],
-              [-74.040288, 40.700644],
-              [-74.03995, 40.700891],
-              [-74.039453, 40.700533],
-              [-74.039383, 40.700578],
-              [-74.037711, 40.699344],
-              [-74.038098, 40.69904],
-              [-74.03823, 40.698369],
-              [-74.039, 40.69837],
-              [-74.03934, 40.698116],
-              [-74.041243, 40.699537],
-              [-74.041748, 40.699148],
-              [-74.039912, 40.697702],
-              [-74.041661, 40.696453],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 98,
-        name: "Financial District",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-74.013754, 40.71369],
-              [-74.008614, 40.711372],
-              [-74.005769, 40.712007],
-              [-74.00536, 40.712151],
-              [-74.003662, 40.711487],
-              [-74.003355, 40.71123],
-              [-74.003337, 40.711236],
-              [-73.999102, 40.707968],
-              [-73.999523, 40.707763],
-              [-73.999491, 40.707722],
-              [-73.999538, 40.707693],
-              [-73.999551, 40.707694],
-              [-73.999609, 40.70776],
-              [-73.999949, 40.707591],
-              [-73.999945, 40.707583],
-              [-73.999949, 40.707559],
-              [-73.999957, 40.707531],
-              [-73.999972, 40.707506],
-              [-74.000234, 40.707347],
-              [-74.001187, 40.70686],
-              [-74.001131, 40.706799],
-              [-74.001232, 40.70674],
-              [-74.00116, 40.706664],
-              [-74.001175, 40.706649],
-              [-74.001169, 40.706639],
-              [-74.001176, 40.706635],
-              [-74.001176, 40.706628],
-              [-74.000983, 40.706417],
-              [-74.001407, 40.706179],
-              [-74.000731, 40.705568],
-              [-74.000583, 40.705434],
-              [-74.000592, 40.70541],
-              [-74.001437, 40.704872],
-              [-74.001461, 40.704874],
-              [-74.001835, 40.705212],
-              [-74.002061, 40.705417],
-              [-74.002158, 40.705357],
-              [-74.00257, 40.705524],
-              [-74.002685, 40.705449],
-              [-74.002363, 40.705159],
-              [-74.002416, 40.705129],
-              [-74.001971, 40.704738],
-              [-74.00197, 40.704731],
-              [-74.002172, 40.704607],
-              [-74.002186, 40.704607],
-              [-74.002333, 40.704748],
-              [-74.002334, 40.704743],
-              [-74.002421, 40.70468],
-              [-74.002431, 40.704678],
-              [-74.002488, 40.704727],
-              [-74.002488, 40.704734],
-              [-74.002398, 40.704795],
-              [-74.002387, 40.704796],
-              [-74.003323, 40.705629],
-              [-74.003632, 40.70543],
-              [-74.003608, 40.705408],
-              [-74.003559, 40.705438],
-              [-74.002878, 40.704831],
-              [-74.002972, 40.704766],
-              [-74.003584, 40.705302],
-              [-74.00365, 40.705378],
-              [-74.003622, 40.705399],
-              [-74.003644, 40.705422],
-              [-74.003833, 40.705303],
-              [-74.003948, 40.705403],
-              [-74.004383, 40.705113],
-              [-74.004267, 40.705012],
-              [-74.00459, 40.704788],
-              [-74.003496, 40.703797],
-              [-74.00372, 40.703654],
-              [-74.003985, 40.703896],
-              [-74.004806, 40.704642],
-              [-74.004841, 40.70464],
-              [-74.005188, 40.704421],
-              [-74.005111, 40.704351],
-              [-74.005085, 40.704366],
-              [-74.004926, 40.704212],
-              [-74.005004, 40.704165],
-              [-74.005163, 40.70432],
-              [-74.005139, 40.704335],
-              [-74.005212, 40.704406],
-              [-74.00535, 40.70431],
-              [-74.004181, 40.70325],
-              [-74.004157, 40.703232],
-              [-74.004207, 40.703199],
-              [-74.004118, 40.703112],
-              [-74.004273, 40.703016],
-              [-74.004298, 40.703017],
-              [-74.004326, 40.703045],
-              [-74.004347, 40.703047],
-              [-74.004677, 40.703348],
-              [-74.004802, 40.703484],
-              [-74.004988, 40.703637],
-              [-74.005208, 40.703821],
-              [-74.005396, 40.703942],
-              [-74.005663, 40.704104],
-              [-74.006019, 40.703873],
-              [-74.00614, 40.703979],
-              [-74.006596, 40.703686],
-              [-74.006324, 40.70347],
-              [-74.005484, 40.702714],
-              [-74.005527, 40.702685],
-              [-74.005505, 40.702667],
-              [-74.005471, 40.70269],
-              [-74.005436, 40.70266],
-              [-74.005427, 40.702665],
-              [-74.005304, 40.70256],
-              [-74.005494, 40.702431],
-              [-74.005502, 40.702431],
-              [-74.005614, 40.70253],
-              [-74.005609, 40.702553],
-              [-74.005599, 40.70258],
-              [-74.005592, 40.702599],
-              [-74.005589, 40.702613],
-              [-74.005535, 40.702646],
-              [-74.005558, 40.702664],
-              [-74.005675, 40.702584],
-              [-74.006239, 40.70308],
-              [-74.006494, 40.703301],
-              [-74.006785, 40.703568],
-              [-74.007527, 40.703115],
-              [-74.007659, 40.702909],
-              [-74.007705, 40.702882],
-              [-74.007707, 40.702875],
-              [-74.009128, 40.702008],
-              [-74.009309, 40.701955],
-              [-74.009085, 40.701655],
-              [-74.0087, 40.70114],
-              [-74.007755, 40.701548],
-              [-74.007591, 40.701328],
-              [-74.008536, 40.70092],
-              [-74.008349, 40.700669],
-              [-74.00862, 40.700552],
-              [-74.009342, 40.70152],
-              [-74.009601, 40.701865],
-              [-74.010511, 40.70164],
-              [-74.011167, 40.701477],
-              [-74.011096, 40.701275],
-              [-74.01129, 40.70125],
-              [-74.01121, 40.701045],
-              [-74.011129, 40.701065],
-              [-74.011078, 40.700957],
-              [-74.011067, 40.700915],
-              [-74.011109, 40.700901],
-              [-74.011094, 40.700866],
-              [-74.011093, 40.700851],
-              [-74.011079, 40.700852],
-              [-74.011056, 40.700812],
-              [-74.011059, 40.700803],
-              [-74.011037, 40.700757],
-              [-74.011015, 40.700741],
-              [-74.011029, 40.700731],
-              [-74.01102, 40.700712],
-              [-74.011007, 40.700711],
-              [-74.010993, 40.700691],
-              [-74.011009, 40.700684],
-              [-74.011004, 40.700668],
-              [-74.010991, 40.700673],
-              [-74.01097, 40.700646],
-              [-74.010988, 40.700634],
-              [-74.010958, 40.700575],
-              [-74.010936, 40.700576],
-              [-74.010925, 40.700567],
-              [-74.010926, 40.700556],
-              [-74.010947, 40.700538],
-              [-74.010919, 40.700486],
-              [-74.010905, 40.700492],
-              [-74.010885, 40.700461],
-              [-74.010904, 40.700451],
-              [-74.010785, 40.700233],
-              [-74.010815, 40.7002],
-              [-74.010852, 40.700239],
-              [-74.01115, 40.700898],
-              [-74.011358, 40.700852],
-              [-74.011303, 40.700648],
-              [-74.011298, 40.700599],
-              [-74.01131, 40.70058],
-              [-74.011334, 40.700568],
-              [-74.011361, 40.700566],
-              [-74.011382, 40.70057],
-              [-74.011395, 40.700591],
-              [-74.011503, 40.700815],
-              [-74.011729, 40.70077],
-              [-74.011683, 40.700693],
-              [-74.011665, 40.700638],
-              [-74.011631, 40.700555],
-              [-74.01161, 40.700493],
-              [-74.011614, 40.700473],
-              [-74.011631, 40.700458],
-              [-74.011678, 40.700462],
-              [-74.011703, 40.700481],
-              [-74.011726, 40.700479],
-              [-74.011751, 40.700531],
-              [-74.011777, 40.700571],
-              [-74.011826, 40.700666],
-              [-74.011854, 40.700739],
-              [-74.012087, 40.700681],
-              [-74.012069, 40.7006],
-              [-74.011977, 40.700378],
-              [-74.011905, 40.700227],
-              [-74.011857, 40.700107],
-              [-74.011931, 40.700051],
-              [-74.011962, 40.700038],
-              [-74.01204, 40.700048],
-              [-74.012088, 40.700087],
-              [-74.012128, 40.700143],
-              [-74.012158, 40.700211],
-              [-74.012237, 40.700321],
-              [-74.012255, 40.700365],
-              [-74.012392, 40.700621],
-              [-74.012594, 40.700571],
-              [-74.012591, 40.700512],
-              [-74.012555, 40.70036],
-              [-74.012529, 40.700252],
-              [-74.012516, 40.700185],
-              [-74.012515, 40.70015],
-              [-74.012542, 40.700135],
-              [-74.012614, 40.700108],
-              [-74.012648, 40.700115],
-              [-74.012681, 40.700147],
-              [-74.012735, 40.700219],
-              [-74.012796, 40.7003],
-              [-74.012835, 40.700353],
-              [-74.012866, 40.700408],
-              [-74.012905, 40.700497],
-              [-74.012909, 40.700524],
-              [-74.013107, 40.700507],
-              [-74.013112, 40.70036],
-              [-74.013078, 40.700103],
-              [-74.013034, 40.699924],
-              [-74.013025, 40.699883],
-              [-74.013012, 40.699818],
-              [-74.013021, 40.699793],
-              [-74.013044, 40.699775],
-              [-74.013113, 40.699774],
-              [-74.013141, 40.699803],
-              [-74.013154, 40.699845],
-              [-74.013167, 40.699931],
-              [-74.013201, 40.70024],
-              [-74.013273, 40.700497],
-              [-74.013486, 40.700478],
-              [-74.013481, 40.700358],
-              [-74.013483, 40.700263],
-              [-74.013503, 40.700176],
-              [-74.013533, 40.700108],
-              [-74.013587, 40.70008],
-              [-74.01361, 40.700084],
-              [-74.013675, 40.700126],
-              [-74.013643, 40.700162],
-              [-74.013612, 40.700245],
-              [-74.01357, 40.700232],
-              [-74.013561, 40.700295],
-              [-74.013533, 40.700297],
-              [-74.01353, 40.700311],
-              [-74.013561, 40.70031],
-              [-74.01356, 40.700367],
-              [-74.01353, 40.700374],
-              [-74.013531, 40.700382],
-              [-74.013563, 40.70038],
-              [-74.01356, 40.700439],
-              [-74.013524, 40.700442],
-              [-74.013523, 40.70046],
-              [-74.013581, 40.700457],
-              [-74.013585, 40.700502],
-              [-74.013553, 40.700503],
-              [-74.013577, 40.700644],
-              [-74.01351, 40.70065],
-              [-74.013558, 40.70094],
-              [-74.013952, 40.700991],
-              [-74.013959, 40.700966],
-              [-74.014008, 40.701043],
-              [-74.015127, 40.70139],
-              [-74.01563, 40.701862],
-              [-74.016174, 40.702572],
-              [-74.016542, 40.704733],
-              [-74.014262, 40.710928],
-              [-74.013754, 40.71369],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 100,
-        name: "Flatiron District",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.983381, 40.741375],
-              [-73.984742, 40.739503],
-              [-73.984743, 40.739502],
-              [-73.985232, 40.738833],
-              [-73.986875, 40.739522],
-              [-73.987743, 40.738311],
-              [-73.989777, 40.739169],
-              [-73.994211, 40.741013],
-              [-73.991356, 40.744763],
-              [-73.983381, 40.741375],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 117,
-        name: "Governors Island",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-74.023056, 40.682917],
-              [-74.023195, 40.683],
-              [-74.022368, 40.68378],
-              [-74.022759, 40.684284],
-              [-74.023017, 40.684167],
-              [-74.023135, 40.684124],
-              [-74.023257, 40.684085],
-              [-74.023398, 40.684041],
-              [-74.023542, 40.684003],
-              [-74.02355, 40.684003],
-              [-74.023783, 40.683969],
-              [-74.023947, 40.683949],
-              [-74.024215, 40.683944],
-              [-74.024432, 40.683954],
-              [-74.024648, 40.683979],
-              [-74.024842, 40.684013],
-              [-74.024993, 40.684052],
-              [-74.025216, 40.684094],
-              [-74.025468, 40.684153],
-              [-74.025672, 40.684236],
-              [-74.025885, 40.684377],
-              [-74.026047, 40.684514],
-              [-74.026126, 40.684594],
-              [-74.026232, 40.6847],
-              [-74.026334, 40.684802],
-              [-74.026441, 40.684952],
-              [-74.026518, 40.685109],
-              [-74.026576, 40.685242],
-              [-74.026609, 40.685377],
-              [-74.026628, 40.685527],
-              [-74.026622, 40.685655],
-              [-74.026609, 40.685763],
-              [-74.02657, 40.685905],
-              [-74.026496, 40.686055],
-              [-74.026426, 40.686175],
-              [-74.026376, 40.686231],
-              [-74.026313, 40.686295],
-              [-74.02622, 40.686385],
-              [-74.024564, 40.687965],
-              [-74.024339, 40.688166],
-              [-74.020205, 40.692034],
-              [-74.019856, 40.692906],
-              [-74.019794, 40.693046],
-              [-74.019768, 40.693081],
-              [-74.019733, 40.693114],
-              [-74.019693, 40.693144],
-              [-74.019647, 40.693169],
-              [-74.019598, 40.693189],
-              [-74.019545, 40.693203],
-              [-74.019436, 40.693221],
-              [-74.019344, 40.693229],
-              [-74.019251, 40.69323],
-              [-74.019159, 40.693225],
-              [-74.019012, 40.69322],
-              [-74.018866, 40.693218],
-              [-74.018851, 40.693218],
-              [-74.018797, 40.693218],
-              [-74.018192, 40.693255],
-              [-74.01796, 40.693269],
-              [-74.017359, 40.693306],
-              [-74.016748, 40.693343],
-              [-74.015404, 40.693066],
-              [-74.015355, 40.693225],
-              [-74.015409, 40.693239],
-              [-74.015389, 40.693298],
-              [-74.015175, 40.693257],
-              [-74.015189, 40.693199],
-              [-74.015255, 40.693207],
-              [-74.015309, 40.693047],
-              [-74.01526, 40.693037],
-              [-74.01512, 40.693188],
-              [-74.015162, 40.693208],
-              [-74.015124, 40.693258],
-              [-74.014931, 40.693156],
-              [-74.014979, 40.693116],
-              [-74.015029, 40.693146],
-              [-74.015156, 40.693016],
-              [-74.014221, 40.692823],
-              [-74.013478, 40.692138],
-              [-74.013472, 40.692132],
-              [-74.013464, 40.692126],
-              [-74.013456, 40.692121],
-              [-74.013446, 40.692118],
-              [-74.013436, 40.692115],
-              [-74.013425, 40.692114],
-              [-74.013414, 40.692115],
-              [-74.013404, 40.692116],
-              [-74.012503, 40.692304],
-              [-74.012523, 40.692339],
-              [-74.012401, 40.692505],
-              [-74.012392, 40.692517],
-              [-74.012373, 40.692534],
-              [-74.01235, 40.692548],
-              [-74.012325, 40.69256],
-              [-74.012297, 40.692568],
-              [-74.012104, 40.692609],
-              [-74.012044, 40.692602],
-              [-74.011849, 40.692518],
-              [-74.011812, 40.692479],
-              [-74.011751, 40.692318],
-              [-74.011757, 40.692264],
-              [-74.011865, 40.692118],
-              [-74.011903, 40.692091],
-              [-74.012149, 40.692038],
-              [-74.012201, 40.692045],
-              [-74.012324, 40.692091],
-              [-74.01239, 40.692123],
-              [-74.012409, 40.692134],
-              [-74.012429, 40.692154],
-              [-74.012461, 40.692212],
-              [-74.012461, 40.692213],
-              [-74.013285, 40.692047],
-              [-74.013303, 40.692043],
-              [-74.013311, 40.692039],
-              [-74.013318, 40.692036],
-              [-74.013324, 40.692032],
-              [-74.013326, 40.69203],
-              [-74.013328, 40.692027],
-              [-74.013331, 40.692024],
-              [-74.013333, 40.69202],
-              [-74.013335, 40.692012],
-              [-74.013335, 40.692008],
-              [-74.013335, 40.692004],
-              [-74.013334, 40.692001],
-              [-74.013325, 40.691985],
-              [-74.013316, 40.691972],
-              [-74.013255, 40.691917],
-              [-74.013187, 40.691958],
-              [-74.012912, 40.691691],
-              [-74.012928, 40.691676],
-              [-74.01286, 40.691603],
-              [-74.012897, 40.691577],
-              [-74.012816, 40.6915],
-              [-74.012754, 40.691476],
-              [-74.012651, 40.691399],
-              [-74.012625, 40.691409],
-              [-74.012609, 40.691364],
-              [-74.012584, 40.691371],
-              [-74.012584, 40.691383],
-              [-74.012545, 40.691392],
-              [-74.012555, 40.691423],
-              [-74.012532, 40.691429],
-              [-74.01252, 40.691397],
-              [-74.012034, 40.691503],
-              [-74.012017, 40.691458],
-              [-74.012598, 40.691334],
-              [-74.012528, 40.691146],
-              [-74.012184, 40.691136],
-              [-74.012187, 40.691078],
-              [-74.011826, 40.691065],
-              [-74.011828, 40.691011],
-              [-74.012195, 40.691029],
-              [-74.01221, 40.690792],
-              [-74.012084, 40.690781],
-              [-74.012082, 40.69076],
-              [-74.012053, 40.690739],
-              [-74.011999, 40.69075],
-              [-74.011988, 40.690772],
-              [-74.011972, 40.690771],
-              [-74.011993, 40.690651],
-              [-74.011981, 40.690606],
-              [-74.01205, 40.690611],
-              [-74.012057, 40.690554],
-              [-74.012099, 40.690547],
-              [-74.012108, 40.690529],
-              [-74.012128, 40.690529],
-              [-74.012142, 40.690454],
-              [-74.012231, 40.690461],
-              [-74.012301, 40.689366],
-              [-74.012307, 40.68928],
-              [-74.013169, 40.688155],
-              [-74.013169, 40.688155],
-              [-74.013406, 40.68803],
-              [-74.014881, 40.687255],
-              [-74.016047, 40.687338],
-              [-74.01705, 40.686876],
-              [-74.017025, 40.686848],
-              [-74.016716, 40.686446],
-              [-74.016647, 40.686416],
-              [-74.015152, 40.686525],
-              [-74.015144, 40.686437],
-              [-74.01662, 40.68632],
-              [-74.016674, 40.686305],
-              [-74.017273, 40.685691],
-              [-74.01753, 40.685363],
-              [-74.017623, 40.68541],
-              [-74.017361, 40.685746],
-              [-74.016781, 40.686349],
-              [-74.016773, 40.686399],
-              [-74.017101, 40.686814],
-              [-74.017121, 40.686843],
-              [-74.020544, 40.685272],
-              [-74.020536, 40.685262],
-              [-74.020207, 40.684834],
-              [-74.020163, 40.684833],
-              [-74.01946, 40.685155],
-              [-74.019398, 40.685087],
-              [-74.020953, 40.684384],
-              [-74.021009, 40.684454],
-              [-74.020278, 40.684809],
-              [-74.020605, 40.68523],
-              [-74.020613, 40.68524],
-              [-74.022508, 40.684398],
-              [-74.022594, 40.68436],
-              [-74.022133, 40.683768],
-              [-74.023056, 40.682917],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 119,
-        name: "Gramercy",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.978527, 40.736855],
-              [-73.982022, 40.732012],
-              [-73.982553, 40.731375],
-              [-73.982554, 40.731375],
-              [-73.990704, 40.734782],
-              [-73.991725, 40.735245],
-              [-73.990371, 40.737098],
-              [-73.989777, 40.739169],
-              [-73.987743, 40.738311],
-              [-73.986875, 40.739522],
-              [-73.985232, 40.738833],
-              [-73.984743, 40.739502],
-              [-73.984742, 40.739503],
-              [-73.97852, 40.736858],
-              [-73.978527, 40.736856],
-              [-73.978527, 40.736855],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 127,
-        name: "Greenwich Village",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.99146, 40.731738],
-              [-73.992854, 40.730083],
-              [-73.996772, 40.725432],
-              [-73.997635, 40.725786],
-              [-74.002856, 40.728347],
-              [-74.001413, 40.731065],
-              [-73.996797, 40.737364],
-              [-73.991725, 40.735245],
-              [-73.990704, 40.734782],
-              [-73.99146, 40.731738],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 129,
-        name: "Harlem",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.934568, 40.828149],
-              [-73.934424, 40.827043],
-              [-73.934346, 40.825777],
-              [-73.934326, 40.825455],
-              [-73.934129, 40.822262],
-              [-73.934104, 40.821855],
-              [-73.933972, 40.81973],
-              [-73.93383, 40.81972],
-              [-73.93383, 40.819521],
-              [-73.934289, 40.819569],
-              [-73.933992, 40.817786],
-              [-73.942754, 40.805744],
-              [-73.941166, 40.805063],
-              [-73.942993, 40.802561],
-              [-73.944597, 40.803228],
-              [-73.949073, 40.797094],
-              [-73.949228, 40.797149],
-              [-73.949374, 40.797158],
-              [-73.949657, 40.797069],
-              [-73.95785, 40.800528],
-              [-73.958074, 40.800359],
-              [-73.958358, 40.800369],
-              [-73.958533, 40.800491],
-              [-73.958509, 40.800682],
-              [-73.959647, 40.801156],
-              [-73.959642, 40.801163],
-              [-73.958249, 40.803107],
-              [-73.958181, 40.805565],
-              [-73.954288, 40.810992],
-              [-73.961038, 40.818704],
-              [-73.96179, 40.818025],
-              [-73.962054, 40.818095],
-              [-73.962029, 40.818123],
-              [-73.961901, 40.818265],
-              [-73.961055, 40.819208],
-              [-73.960839, 40.819447],
-              [-73.960543, 40.819777],
-              [-73.960242, 40.820112],
-              [-73.959999, 40.820383],
-              [-73.959949, 40.82036],
-              [-73.959815, 40.820298],
-              [-73.959283, 40.820995],
-              [-73.959131, 40.821169],
-              [-73.959263, 40.821236],
-              [-73.959303, 40.82119],
-              [-73.959426, 40.821252],
-              [-73.959147, 40.821571],
-              [-73.959033, 40.821514],
-              [-73.959235, 40.821278],
-              [-73.959147, 40.821235],
-              [-73.95909, 40.821206],
-              [-73.958819, 40.821519],
-              [-73.958887, 40.821547],
-              [-73.958942, 40.821569],
-              [-73.958745, 40.821806],
-              [-73.958964, 40.822374],
-              [-73.959, 40.82241],
-              [-73.959113, 40.822455],
-              [-73.95923, 40.822287],
-              [-73.95991, 40.822571],
-              [-73.959603, 40.822995],
-              [-73.958926, 40.822712],
-              [-73.95905, 40.822541],
-              [-73.958917, 40.822486],
-              [-73.958867, 40.82245],
-              [-73.958838, 40.822405],
-              [-73.958646, 40.821904],
-              [-73.958527, 40.821844],
-              [-73.9582, 40.822283],
-              [-73.958168, 40.822458],
-              [-73.958075, 40.822611],
-              [-73.957939, 40.822774],
-              [-73.958228, 40.822888],
-              [-73.958143, 40.823021],
-              [-73.959483, 40.823579],
-              [-73.959538, 40.823629],
-              [-73.959556, 40.823694],
-              [-73.959016, 40.824504],
-              [-73.957259, 40.827117],
-              [-73.956949, 40.827001],
-              [-73.956318, 40.827944],
-              [-73.955039, 40.827413],
-              [-73.954706, 40.827906],
-              [-73.954697, 40.827918],
-              [-73.954545, 40.827863],
-              [-73.954542, 40.827861],
-              [-73.954526, 40.827884],
-              [-73.952928, 40.830158],
-              [-73.952345, 40.83102],
-              [-73.951732, 40.831859],
-              [-73.951316, 40.832451],
-              [-73.951059, 40.832898],
-              [-73.950936, 40.833294],
-              [-73.950664, 40.833636],
-              [-73.950423, 40.834047],
-              [-73.95022, 40.834305],
-              [-73.949342, 40.833982],
-              [-73.948905, 40.834066],
-              [-73.940346, 40.830458],
-              [-73.934568, 40.828149],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 131,
-        name: "Hell's Kitchen",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.990724, 40.75595],
-              [-73.990726, 40.755951],
-              [-73.993464, 40.75219],
-              [-74.004927, 40.757023],
-              [-74.005567, 40.757118],
-              [-74.005419, 40.757323],
-              [-74.005212, 40.757232],
-              [-74.004804, 40.75781],
-              [-74.005279, 40.758029],
-              [-74.005235, 40.758087],
-              [-74.005031, 40.758357],
-              [-74.006964, 40.75916],
-              [-74.006994, 40.759176],
-              [-74.00701, 40.759204],
-              [-74.007004, 40.759232],
-              [-74.00653, 40.759875],
-              [-74.006507, 40.759892],
-              [-74.006476, 40.759894],
-              [-74.006444, 40.75989],
-              [-74.006315, 40.759843],
-              [-74.004505, 40.759085],
-              [-74.004307, 40.759361],
-              [-74.004206, 40.759502],
-              [-74.003907, 40.759376],
-              [-74.003807, 40.759511],
-              [-74.004155, 40.75966],
-              [-74.004247, 40.759699],
-              [-74.004165, 40.759813],
-              [-74.004149, 40.759836],
-              [-74.004245, 40.759887],
-              [-74.00426, 40.759875],
-              [-74.004319, 40.759911],
-              [-74.004269, 40.759959],
-              [-74.004211, 40.759925],
-              [-74.004223, 40.759913],
-              [-74.004132, 40.759862],
-              [-74.004112, 40.759893],
-              [-74.003669, 40.759706],
-              [-74.003573, 40.759846],
-              [-74.003573, 40.759847],
-              [-74.003687, 40.759897],
-              [-74.003929, 40.760003],
-              [-74.003927, 40.760005],
-              [-74.003882, 40.760066],
-              [-74.003171, 40.761027],
-              [-74.003133, 40.761078],
-              [-74.002787, 40.760929],
-              [-74.002669, 40.76109],
-              [-74.002335, 40.761544],
-              [-74.004545, 40.762493],
-              [-74.004558, 40.762508],
-              [-74.004556, 40.762524],
-              [-74.00445, 40.762665],
-              [-74.004427, 40.762676],
-              [-74.0044, 40.762671],
-              [-74.002201, 40.761728],
-              [-74.001533, 40.762645],
-              [-74.003757, 40.763582],
-              [-74.003612, 40.763777],
-              [-74.00334, 40.763664],
-              [-74.001386, 40.76285],
-              [-74.000973, 40.76342],
-              [-74.00315, 40.764333],
-              [-74.002948, 40.764617],
-              [-74.003062, 40.764666],
-              [-74.003135, 40.76465],
-              [-74.003162, 40.764613],
-              [-74.003223, 40.764638],
-              [-74.003156, 40.764731],
-              [-74.003147, 40.764736],
-              [-74.003136, 40.764741],
-              [-74.003125, 40.764744],
-              [-74.003113, 40.764746],
-              [-74.003101, 40.764746],
-              [-74.003089, 40.764744],
-              [-73.999927, 40.763407],
-              [-73.999639, 40.763569],
-              [-73.999623, 40.763559],
-              [-73.999151, 40.764205],
-              [-73.999314, 40.764272],
-              [-73.99952, 40.763974],
-              [-74.001411, 40.764772],
-              [-74.001498, 40.764675],
-              [-74.001939, 40.764867],
-              [-74.00186, 40.764962],
-              [-74.002457, 40.765214],
-              [-74.002221, 40.765501],
-              [-74.002274, 40.765525],
-              [-74.002288, 40.765562],
-              [-74.002057, 40.765877],
-              [-74.002025, 40.765893],
-              [-73.998885, 40.764569],
-              [-73.998576, 40.764999],
-              [-73.9986, 40.765073],
-              [-73.998635, 40.765086],
-              [-73.998665, 40.765104],
-              [-73.998686, 40.765137],
-              [-73.998696, 40.765175],
-              [-73.998689, 40.765209],
-              [-73.998668, 40.765237],
-              [-73.998634, 40.76526],
-              [-73.998589, 40.765276],
-              [-73.998483, 40.765298],
-              [-73.998438, 40.765316],
-              [-73.998402, 40.765351],
-              [-73.998291, 40.765521],
-              [-73.998271, 40.765552],
-              [-73.998282, 40.765568],
-              [-74.001542, 40.766945],
-              [-74.001573, 40.766977],
-              [-74.001562, 40.766996],
-              [-74.001394, 40.76723],
-              [-74.001364, 40.767249],
-              [-74.001315, 40.767251],
-              [-73.998157, 40.765894],
-              [-73.998148, 40.765906],
-              [-73.998067, 40.765872],
-              [-73.997382, 40.766818],
-              [-73.997378, 40.766823],
-              [-74.000628, 40.768203],
-              [-74.00065, 40.76822],
-              [-74.00065, 40.768245],
-              [-74.000455, 40.768504],
-              [-74.000418, 40.768512],
-              [-74.000386, 40.768504],
-              [-73.997156, 40.767129],
-              [-73.996457, 40.768099],
-              [-73.999728, 40.769474],
-              [-73.999736, 40.769506],
-              [-73.999555, 40.769751],
-              [-73.999516, 40.769768],
-              [-73.999473, 40.769764],
-              [-73.997072, 40.768733],
-              [-73.997026, 40.768796],
-              [-73.99715, 40.768847],
-              [-73.997145, 40.768853],
-              [-73.997011, 40.769029],
-              [-73.996893, 40.76898],
-              [-73.996571, 40.769425],
-              [-73.999027, 40.770463],
-              [-73.998762, 40.770825],
-              [-73.996313, 40.769799],
-              [-73.99588, 40.770401],
-              [-73.996017, 40.770455],
-              [-73.995901, 40.770613],
-              [-73.995767, 40.770555],
-              [-73.995425, 40.771032],
-              [-73.995756, 40.771161],
-              [-73.995709, 40.771226],
-              [-73.995798, 40.771283],
-              [-73.995817, 40.771257],
-              [-73.995913, 40.771298],
-              [-73.995928, 40.771275],
-              [-73.996002, 40.771307],
-              [-73.9959, 40.77145],
-              [-73.995893, 40.771447],
-              [-73.995854, 40.771499],
-              [-73.995791, 40.771476],
-              [-73.995828, 40.771422],
-              [-73.99573, 40.771379],
-              [-73.995751, 40.771354],
-              [-73.99563, 40.771325],
-              [-73.995623, 40.771337],
-              [-73.995306, 40.771196],
-              [-73.995124, 40.771451],
-              [-73.99499, 40.771441],
-              [-73.994957, 40.771449],
-              [-73.994935, 40.771468],
-              [-73.994924, 40.771486],
-              [-73.994926, 40.771516],
-              [-73.995009, 40.771607],
-              [-73.994978, 40.771645],
-              [-73.994977, 40.771659],
-              [-73.997112, 40.772553],
-              [-73.996909, 40.772833],
-              [-73.994774, 40.77193],
-              [-73.994532, 40.772259],
-              [-73.994445, 40.772377],
-              [-73.995455, 40.772803],
-              [-73.995545, 40.772675],
-              [-73.996449, 40.773057],
-              [-73.996356, 40.773179],
-              [-73.99623, 40.773129],
-              [-73.996143, 40.773248],
-              [-73.996562, 40.773411],
-              [-73.996685, 40.773504],
-              [-73.996677, 40.77351],
-              [-73.994342, 40.772525],
-              [-73.99425, 40.772526],
-              [-73.994163, 40.772487],
-              [-73.994159, 40.772493],
-              [-73.99365, 40.772146],
-              [-73.982389, 40.767411],
-              [-73.984726, 40.764167],
-              [-73.990724, 40.75595],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 142,
-        name: "Inwood",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.922298, 40.855599],
-              [-73.922531, 40.85568],
-              [-73.922906, 40.855912],
-              [-73.92293, 40.858941],
-              [-73.92471, 40.861552],
-              [-73.927091, 40.85976],
-              [-73.927625, 40.859012],
-              [-73.930122, 40.858744],
-              [-73.931313, 40.859328],
-              [-73.933053, 40.859828],
-              [-73.935532, 40.858918],
-              [-73.937789, 40.859812],
-              [-73.937815, 40.859834],
-              [-73.937704, 40.859936],
-              [-73.937529, 40.860097],
-              [-73.937581, 40.860203],
-              [-73.937177, 40.860709],
-              [-73.937042, 40.861058],
-              [-73.936969, 40.861143],
-              [-73.936811, 40.861258],
-              [-73.936803, 40.861279],
-              [-73.936819, 40.861316],
-              [-73.93681, 40.861363],
-              [-73.936476, 40.861927],
-              [-73.936228, 40.862239],
-              [-73.935529, 40.863124],
-              [-73.934434, 40.864504],
-              [-73.934356, 40.864733],
-              [-73.934017, 40.865036],
-              [-73.934016, 40.865196],
-              [-73.93395, 40.865347],
-              [-73.933402, 40.866136],
-              [-73.933323, 40.866238],
-              [-73.933257, 40.866242],
-              [-73.932663, 40.867237],
-              [-73.932756, 40.867264],
-              [-73.932763, 40.867252],
-              [-73.932902, 40.867293],
-              [-73.93284, 40.867415],
-              [-73.932646, 40.867358],
-              [-73.93266, 40.86733],
-              [-73.932598, 40.867311],
-              [-73.932548, 40.867414],
-              [-73.932519, 40.867411],
-              [-73.932425, 40.867568],
-              [-73.932373, 40.868063],
-              [-73.932274, 40.868477],
-              [-73.932281, 40.868533],
-              [-73.932268, 40.868562],
-              [-73.932323, 40.868578],
-              [-73.932315, 40.868604],
-              [-73.932492, 40.868678],
-              [-73.932625, 40.868456],
-              [-73.932651, 40.868465],
-              [-73.932545, 40.86864],
-              [-73.932568, 40.868648],
-              [-73.932524, 40.868722],
-              [-73.932501, 40.868714],
-              [-73.932506, 40.868706],
-              [-73.932306, 40.868629],
-              [-73.932298, 40.868647],
-              [-73.93225, 40.86864],
-              [-73.932254, 40.868683],
-              [-73.932278, 40.868711],
-              [-73.932314, 40.868924],
-              [-73.93236, 40.86895],
-              [-73.932537, 40.868968],
-              [-73.932569, 40.868979],
-              [-73.932568, 40.869044],
-              [-73.932505, 40.86913],
-              [-73.93239, 40.869187],
-              [-73.932293, 40.8692],
-              [-73.932189, 40.869367],
-              [-73.932469, 40.869476],
-              [-73.932354, 40.869755],
-              [-73.932266, 40.869735],
-              [-73.932359, 40.869507],
-              [-73.932133, 40.869416],
-              [-73.932053, 40.869601],
-              [-73.932096, 40.86976],
-              [-73.932199, 40.869845],
-              [-73.932206, 40.870096],
-              [-73.932032, 40.870757],
-              [-73.931914, 40.871129],
-              [-73.931377, 40.872109],
-              [-73.931063, 40.872667],
-              [-73.929977, 40.874339],
-              [-73.929013, 40.875739],
-              [-73.928146, 40.876523],
-              [-73.927966, 40.876618],
-              [-73.927531, 40.876711],
-              [-73.92708, 40.877075],
-              [-73.926531, 40.877584],
-              [-73.926482, 40.877618],
-              [-73.926416, 40.87761],
-              [-73.926406, 40.877621],
-              [-73.926297, 40.877565],
-              [-73.92631, 40.877551],
-              [-73.926303, 40.877537],
-              [-73.926297, 40.877524],
-              [-73.926295, 40.877506],
-              [-73.926295, 40.877488],
-              [-73.9263, 40.877471],
-              [-73.926309, 40.877455],
-              [-73.92632, 40.87744],
-              [-73.926335, 40.877426],
-              [-73.926353, 40.877414],
-              [-73.926373, 40.877405],
-              [-73.926394, 40.877399],
-              [-73.926441, 40.877332],
-              [-73.926458, 40.877258],
-              [-73.926414, 40.877181],
-              [-73.926341, 40.877119],
-              [-73.926193, 40.877064],
-              [-73.925553, 40.877103],
-              [-73.925176, 40.877255],
-              [-73.925169, 40.877283],
-              [-73.925158, 40.877311],
-              [-73.925138, 40.877342],
-              [-73.925113, 40.877372],
-              [-73.925082, 40.877398],
-              [-73.925027, 40.877372],
-              [-73.924962, 40.877394],
-              [-73.924896, 40.877411],
-              [-73.924742, 40.877432],
-              [-73.92457, 40.877429],
-              [-73.923295, 40.877264],
-              [-73.922447, 40.876807],
-              [-73.922398, 40.87678],
-              [-73.922388, 40.876763],
-              [-73.922377, 40.876759],
-              [-73.922375, 40.876557],
-              [-73.922381, 40.876366],
-              [-73.922435, 40.875912],
-              [-73.922445, 40.875692],
-              [-73.922668, 40.874982],
-              [-73.922667, 40.874968],
-              [-73.922614, 40.874407],
-              [-73.922177, 40.873675],
-              [-73.921255, 40.873082],
-              [-73.920556, 40.872999],
-              [-73.920134, 40.873193],
-              [-73.919927, 40.873595],
-              [-73.919905, 40.873611],
-              [-73.919793, 40.873607],
-              [-73.919793, 40.873646],
-              [-73.9199, 40.873649],
-              [-73.919925, 40.873654],
-              [-73.919958, 40.873679],
-              [-73.92013, 40.873923],
-              [-73.920145, 40.873905],
-              [-73.920241, 40.873963],
-              [-73.920223, 40.873979],
-              [-73.920441, 40.874107],
-              [-73.9205, 40.87416],
-              [-73.920544, 40.874145],
-              [-73.920606, 40.874207],
-              [-73.920627, 40.874235],
-              [-73.92077, 40.874177],
-              [-73.920788, 40.874179],
-              [-73.920803, 40.874194],
-              [-73.920808, 40.874215],
-              [-73.920661, 40.874279],
-              [-73.920686, 40.874323],
-              [-73.920711, 40.874387],
-              [-73.9207, 40.874392],
-              [-73.920716, 40.874406],
-              [-73.920725, 40.874431],
-              [-73.920758, 40.874466],
-              [-73.92114, 40.875048],
-              [-73.921175, 40.875106],
-              [-73.921182, 40.875138],
-              [-73.92117, 40.875183],
-              [-73.92116, 40.875196],
-              [-73.921127, 40.875235],
-              [-73.921111, 40.875238],
-              [-73.921104, 40.875253],
-              [-73.921113, 40.875267],
-              [-73.921107, 40.875276],
-              [-73.921105, 40.875285],
-              [-73.92108, 40.875293],
-              [-73.921062, 40.875309],
-              [-73.921057, 40.875319],
-              [-73.921036, 40.875344],
-              [-73.921024, 40.875368],
-              [-73.921011, 40.875377],
-              [-73.92099, 40.875385],
-              [-73.920978, 40.8754],
-              [-73.920969, 40.875419],
-              [-73.920955, 40.875429],
-              [-73.92093, 40.875437],
-              [-73.920929, 40.875446],
-              [-73.920918, 40.875453],
-              [-73.920913, 40.875465],
-              [-73.920903, 40.875466],
-              [-73.920894, 40.875458],
-              [-73.920876, 40.875464],
-              [-73.92081, 40.875454],
-              [-73.920665, 40.875394],
-              [-73.920601, 40.875411],
-              [-73.920525, 40.875409],
-              [-73.92048, 40.875375],
-              [-73.920481, 40.875339],
-              [-73.919701, 40.874994],
-              [-73.919554, 40.874952],
-              [-73.919411, 40.874903],
-              [-73.919271, 40.874849],
-              [-73.919135, 40.874789],
-              [-73.919003, 40.874723],
-              [-73.918877, 40.874653],
-              [-73.918862, 40.874643],
-              [-73.91885, 40.874622],
-              [-73.918855, 40.874593],
-              [-73.918944, 40.874521],
-              [-73.919064, 40.8744],
-              [-73.919114, 40.874334],
-              [-73.919193, 40.874195],
-              [-73.919242, 40.874048],
-              [-73.91927, 40.87381],
-              [-73.919271, 40.87369],
-              [-73.919271, 40.873677],
-              [-73.919277, 40.873664],
-              [-73.919288, 40.873654],
-              [-73.919295, 40.87365],
-              [-73.919281, 40.873633],
-              [-73.919265, 40.873641],
-              [-73.919247, 40.873645],
-              [-73.919223, 40.873641],
-              [-73.919201, 40.873634],
-              [-73.91918, 40.873625],
-              [-73.919169, 40.87362],
-              [-73.91916, 40.873613],
-              [-73.919153, 40.873605],
-              [-73.919147, 40.873596],
-              [-73.919143, 40.873587],
-              [-73.919141, 40.873577],
-              [-73.919142, 40.873567],
-              [-73.919141, 40.873546],
-              [-73.919136, 40.873526],
-              [-73.919127, 40.873506],
-              [-73.919114, 40.873487],
-              [-73.919098, 40.873471],
-              [-73.919098, 40.873469],
-              [-73.919096, 40.873466],
-              [-73.919092, 40.873465],
-              [-73.91909, 40.873464],
-              [-73.919087, 40.873465],
-              [-73.919086, 40.873465],
-              [-73.919085, 40.873466],
-              [-73.919082, 40.873467],
-              [-73.919078, 40.873468],
-              [-73.919075, 40.873469],
-              [-73.919071, 40.873469],
-              [-73.919068, 40.873469],
-              [-73.919064, 40.873468],
-              [-73.919061, 40.873467],
-              [-73.919058, 40.873466],
-              [-73.919059, 40.873464],
-              [-73.919059, 40.873463],
-              [-73.919058, 40.873462],
-              [-73.919058, 40.873461],
-              [-73.919057, 40.87346],
-              [-73.919056, 40.873459],
-              [-73.919055, 40.873458],
-              [-73.919054, 40.873458],
-              [-73.919052, 40.873457],
-              [-73.919051, 40.873457],
-              [-73.919049, 40.873457],
-              [-73.919048, 40.873457],
-              [-73.919046, 40.873457],
-              [-73.919045, 40.873458],
-              [-73.919044, 40.873459],
-              [-73.919036, 40.873458],
-              [-73.919029, 40.873457],
-              [-73.919022, 40.873455],
-              [-73.919016, 40.873452],
-              [-73.91901, 40.873449],
-              [-73.919005, 40.873444],
-              [-73.919, 40.873438],
-              [-73.918994, 40.873432],
-              [-73.918986, 40.873428],
-              [-73.918978, 40.873424],
-              [-73.918969, 40.873421],
-              [-73.91896, 40.873419],
-              [-73.91895, 40.873418],
-              [-73.918944, 40.873417],
-              [-73.918937, 40.873415],
-              [-73.918931, 40.873413],
-              [-73.918926, 40.87341],
-              [-73.918921, 40.873406],
-              [-73.918918, 40.873402],
-              [-73.918916, 40.873394],
-              [-73.918917, 40.873387],
-              [-73.918918, 40.873379],
-              [-73.918922, 40.873372],
-              [-73.918926, 40.873366],
-              [-73.918932, 40.87336],
-              [-73.918939, 40.873355],
-              [-73.918947, 40.87335],
-              [-73.918951, 40.873349],
-              [-73.918955, 40.873346],
-              [-73.918957, 40.873342],
-              [-73.918957, 40.873339],
-              [-73.918954, 40.873335],
-              [-73.918948, 40.873333],
-              [-73.918943, 40.873331],
-              [-73.918938, 40.873328],
-              [-73.918934, 40.873324],
-              [-73.918931, 40.873321],
-              [-73.918929, 40.873317],
-              [-73.918916, 40.873324],
-              [-73.918897, 40.873327],
-              [-73.918871, 40.873328],
-              [-73.918864, 40.873334],
-              [-73.91886, 40.873341],
-              [-73.918848, 40.873344],
-              [-73.918839, 40.873346],
-              [-73.918829, 40.873345],
-              [-73.918821, 40.873346],
-              [-73.918811, 40.873348],
-              [-73.9188, 40.873349],
-              [-73.918788, 40.873347],
-              [-73.918776, 40.873344],
-              [-73.918766, 40.873338],
-              [-73.918758, 40.873328],
-              [-73.918748, 40.87332],
-              [-73.918725, 40.873312],
-              [-73.91872, 40.873307],
-              [-73.918713, 40.873305],
-              [-73.918698, 40.873278],
-              [-73.918699, 40.873271],
-              [-73.918723, 40.873249],
-              [-73.918713, 40.873246],
-              [-73.918708, 40.873248],
-              [-73.918696, 40.873252],
-              [-73.918672, 40.873254],
-              [-73.91866, 40.873261],
-              [-73.918649, 40.873279],
-              [-73.918632, 40.873284],
-              [-73.918617, 40.873286],
-              [-73.918603, 40.873286],
-              [-73.918587, 40.873278],
-              [-73.91857, 40.873263],
-              [-73.918565, 40.873251],
-              [-73.918566, 40.873238],
-              [-73.918551, 40.873222],
-              [-73.918531, 40.873213],
-              [-73.918516, 40.873214],
-              [-73.918508, 40.873207],
-              [-73.9185, 40.873192],
-              [-73.918481, 40.873196],
-              [-73.918409, 40.873197],
-              [-73.918391, 40.873209],
-              [-73.918397, 40.873232],
-              [-73.918417, 40.87327],
-              [-73.918532, 40.873325],
-              [-73.918594, 40.873357],
-              [-73.918613, 40.873371],
-              [-73.918629, 40.873405],
-              [-73.918628, 40.87344],
-              [-73.918612, 40.873473],
-              [-73.918597, 40.873495],
-              [-73.918573, 40.873496],
-              [-73.918561, 40.87351],
-              [-73.918496, 40.873548],
-              [-73.918469, 40.873571],
-              [-73.918454, 40.873568],
-              [-73.918439, 40.873554],
-              [-73.918412, 40.873551],
-              [-73.918401, 40.873563],
-              [-73.918425, 40.873586],
-              [-73.918425, 40.873602],
-              [-73.918411, 40.873637],
-              [-73.918328, 40.873694],
-              [-73.918295, 40.873715],
-              [-73.918246, 40.873737],
-              [-73.918199, 40.873753],
-              [-73.918179, 40.873769],
-              [-73.918162, 40.873791],
-              [-73.918147, 40.873822],
-              [-73.918104, 40.873847],
-              [-73.918059, 40.87385],
-              [-73.918037, 40.873897],
-              [-73.917941, 40.873983],
-              [-73.917929, 40.873979],
-              [-73.917755, 40.874249],
-              [-73.91777, 40.874254],
-              [-73.917577, 40.874549],
-              [-73.917499, 40.87452],
-              [-73.917691, 40.874225],
-              [-73.917709, 40.874232],
-              [-73.91789, 40.873962],
-              [-73.917885, 40.87396],
-              [-73.917871, 40.873905],
-              [-73.917847, 40.873931],
-              [-73.917814, 40.873946],
-              [-73.917808, 40.873957],
-              [-73.917793, 40.873953],
-              [-73.917778, 40.873963],
-              [-73.917781, 40.87398],
-              [-73.917768, 40.873992],
-              [-73.91774, 40.874007],
-              [-73.917396, 40.87416],
-              [-73.916981, 40.874327],
-              [-73.916713, 40.874355],
-              [-73.91658, 40.874436],
-              [-73.916448, 40.874495],
-              [-73.916343, 40.874524],
-              [-73.915685, 40.874528],
-              [-73.915224, 40.874487],
-              [-73.912729, 40.873807],
-              [-73.912502, 40.873728],
-              [-73.911942, 40.873458],
-              [-73.911578, 40.873272],
-              [-73.911175, 40.872985],
-              [-73.910649, 40.872296],
-              [-73.910541, 40.872058],
-              [-73.910538, 40.872034],
-              [-73.910553, 40.872018],
-              [-73.910518, 40.871906],
-              [-73.910495, 40.871826],
-              [-73.910473, 40.871696],
-              [-73.910461, 40.871642],
-              [-73.910448, 40.871502],
-              [-73.910443, 40.87149],
-              [-73.910426, 40.871485],
-              [-73.910429, 40.871466],
-              [-73.910444, 40.871464],
-              [-73.910434, 40.871381],
-              [-73.910678, 40.870873],
-              [-73.910748, 40.870265],
-              [-73.910853, 40.869649],
-              [-73.910919, 40.869524],
-              [-73.910941, 40.869489],
-              [-73.910947, 40.869452],
-              [-73.910925, 40.869437],
-              [-73.910956, 40.869411],
-              [-73.910987, 40.869405],
-              [-73.911009, 40.869419],
-              [-73.911039, 40.869422],
-              [-73.911058, 40.869409],
-              [-73.911132, 40.869442],
-              [-73.911189, 40.869459],
-              [-73.911256, 40.869487],
-              [-73.911304, 40.86951],
-              [-73.911346, 40.869524],
-              [-73.911382, 40.869523],
-              [-73.91144, 40.869495],
-              [-73.911464, 40.869477],
-              [-73.911474, 40.869455],
-              [-73.911472, 40.869413],
-              [-73.911463, 40.869377],
-              [-73.911424, 40.869364],
-              [-73.911417, 40.869355],
-              [-73.9114, 40.869352],
-              [-73.911394, 40.869344],
-              [-73.911358, 40.869336],
-              [-73.911355, 40.869313],
-              [-73.911122, 40.869214],
-              [-73.911117, 40.869179],
-              [-73.911097, 40.869179],
-              [-73.911031, 40.869155],
-              [-73.911076, 40.869043],
-              [-73.911129, 40.86892],
-              [-73.911149, 40.86888],
-              [-73.9112, 40.868864],
-              [-73.911251, 40.868861],
-              [-73.911283, 40.868848],
-              [-73.911328, 40.868802],
-              [-73.911337, 40.868774],
-              [-73.91133, 40.868759],
-              [-73.911325, 40.868718],
-              [-73.911326, 40.868676],
-              [-73.911364, 40.868583],
-              [-73.911505, 40.868303],
-              [-73.911848, 40.867673],
-              [-73.912732, 40.866605],
-              [-73.912548, 40.866515],
-              [-73.913612, 40.865142],
-              [-73.913919, 40.864746],
-              [-73.914312, 40.864625],
-              [-73.914924, 40.86458],
-              [-73.914929, 40.86457],
-              [-73.91454, 40.864412],
-              [-73.914474, 40.864311],
-              [-73.914411, 40.864174],
-              [-73.915086, 40.864448],
-              [-73.915432, 40.864418],
-              [-73.915429, 40.864408],
-              [-73.915434, 40.864387],
-              [-73.915459, 40.864384],
-              [-73.915453, 40.864334],
-              [-73.915632, 40.864309],
-              [-73.915639, 40.864349],
-              [-73.915664, 40.86435],
-              [-73.915712, 40.864343],
-              [-73.915817, 40.864335],
-              [-73.915899, 40.864308],
-              [-73.915936, 40.864242],
-              [-73.915926, 40.864173],
-              [-73.915871, 40.864129],
-              [-73.91579, 40.864103],
-              [-73.915667, 40.864087],
-              [-73.915584, 40.864071],
-              [-73.915493, 40.864066],
-              [-73.91544, 40.864078],
-              [-73.915384, 40.86408],
-              [-73.915346, 40.864073],
-              [-73.915311, 40.864051],
-              [-73.915298, 40.864035],
-              [-73.915267, 40.86402],
-              [-73.915241, 40.863996],
-              [-73.915216, 40.863976],
-              [-73.915196, 40.863945],
-              [-73.915182, 40.86386],
-              [-73.915175, 40.863708],
-              [-73.915176, 40.86368],
-              [-73.915201, 40.863652],
-              [-73.915254, 40.863618],
-              [-73.915308, 40.86359],
-              [-73.915362, 40.863547],
-              [-73.915477, 40.863346],
-              [-73.915485, 40.863301],
-              [-73.91546, 40.863257],
-              [-73.915451, 40.863233],
-              [-73.915455, 40.863208],
-              [-73.915423, 40.863194],
-              [-73.915411, 40.863195],
-              [-73.915345, 40.863165],
-              [-73.915411, 40.863095],
-              [-73.915387, 40.863084],
-              [-73.91552, 40.86291],
-              [-73.915566, 40.862931],
-              [-73.915576, 40.862916],
-              [-73.915563, 40.862909],
-              [-73.91558, 40.862896],
-              [-73.915589, 40.862901],
-              [-73.9156, 40.862892],
-              [-73.915594, 40.862889],
-              [-73.915604, 40.86288],
-              [-73.915618, 40.862879],
-              [-73.915672, 40.862902],
-              [-73.915682, 40.862894],
-              [-73.91578, 40.862789],
-              [-73.915952, 40.862631],
-              [-73.916105, 40.862488],
-              [-73.916276, 40.862366],
-              [-73.916492, 40.862198],
-              [-73.916795, 40.861978],
-              [-73.917295, 40.861679],
-              [-73.91741, 40.861486],
-              [-73.917481, 40.861399],
-              [-73.917825, 40.86098],
-              [-73.917821, 40.860949],
-              [-73.917841, 40.860931],
-              [-73.917861, 40.860901],
-              [-73.917865, 40.860884],
-              [-73.917877, 40.860884],
-              [-73.917882, 40.860874],
-              [-73.917905, 40.86085],
-              [-73.917902, 40.86084],
-              [-73.917922, 40.860825],
-              [-73.917922, 40.860814],
-              [-73.917884, 40.860802],
-              [-73.918434, 40.860222],
-              [-73.918393, 40.86008],
-              [-73.918423, 40.860039],
-              [-73.918778, 40.859547],
-              [-73.919371, 40.858926],
-              [-73.919502, 40.858791],
-              [-73.919524, 40.858783],
-              [-73.919546, 40.858792],
-              [-73.919547, 40.858799],
-              [-73.92029, 40.859106],
-              [-73.920281, 40.859115],
-              [-73.920361, 40.859149],
-              [-73.920412, 40.859148],
-              [-73.920435, 40.859121],
-              [-73.920445, 40.859124],
-              [-73.920461, 40.859103],
-              [-73.919847, 40.858844],
-              [-73.919772, 40.858802],
-              [-73.919774, 40.858795],
-              [-73.920479, 40.859104],
-              [-73.920431, 40.859155],
-              [-73.920441, 40.859161],
-              [-73.920434, 40.859172],
-              [-73.920509, 40.859205],
-              [-73.920508, 40.859212],
-              [-73.92062, 40.859259],
-              [-73.920653, 40.859263],
-              [-73.920698, 40.859324],
-              [-73.920693, 40.859351],
-              [-73.920748, 40.859398],
-              [-73.920902, 40.859683],
-              [-73.921044, 40.859877],
-              [-73.921271, 40.860188],
-              [-73.921322, 40.860214],
-              [-73.92134, 40.860217],
-              [-73.921346, 40.860213],
-              [-73.921364, 40.860211],
-              [-73.921394, 40.860191],
-              [-73.921459, 40.860163],
-              [-73.921516, 40.860131],
-              [-73.921516, 40.860123],
-              [-73.921543, 40.860099],
-              [-73.921557, 40.860099],
-              [-73.92159, 40.860073],
-              [-73.921661, 40.859891],
-              [-73.921688, 40.85984],
-              [-73.921718, 40.859712],
-              [-73.921701, 40.859694],
-              [-73.92171, 40.859672],
-              [-73.921737, 40.859672],
-              [-73.921738, 40.859647],
-              [-73.921723, 40.859634],
-              [-73.921704, 40.859631],
-              [-73.921686, 40.859636],
-              [-73.921697, 40.859612],
-              [-73.921681, 40.859609],
-              [-73.921665, 40.859592],
-              [-73.921671, 40.859581],
-              [-73.921669, 40.859572],
-              [-73.921648, 40.859571],
-              [-73.921649, 40.859551],
-              [-73.92169, 40.859548],
-              [-73.921731, 40.859556],
-              [-73.92172, 40.859568],
-              [-73.921754, 40.859569],
-              [-73.92174, 40.85958],
-              [-73.921759, 40.859579],
-              [-73.921769, 40.859559],
-              [-73.921803, 40.859559],
-              [-73.921803, 40.859548],
-              [-73.921817, 40.859542],
-              [-73.921816, 40.85956],
-              [-73.921833, 40.859555],
-              [-73.92184, 40.859522],
-              [-73.921826, 40.859508],
-              [-73.921801, 40.859495],
-              [-73.921784, 40.859497],
-              [-73.921784, 40.85948],
-              [-73.92179, 40.859476],
-              [-73.921791, 40.859467],
-              [-73.921783, 40.85946],
-              [-73.921778, 40.859437],
-              [-73.921755, 40.859411],
-              [-73.921745, 40.859412],
-              [-73.921728, 40.859395],
-              [-73.921739, 40.859391],
-              [-73.921742, 40.859378],
-              [-73.921781, 40.859376],
-              [-73.921801, 40.859384],
-              [-73.921814, 40.859395],
-              [-73.921846, 40.859383],
-              [-73.921793, 40.859353],
-              [-73.921806, 40.859329],
-              [-73.921836, 40.859323],
-              [-73.921842, 40.8593],
-              [-73.921857, 40.859291],
-              [-73.921856, 40.859271],
-              [-73.921872, 40.8592],
-              [-73.921866, 40.859181],
-              [-73.921849, 40.859163],
-              [-73.92181, 40.859141],
-              [-73.921829, 40.859129],
-              [-73.921818, 40.859115],
-              [-73.921795, 40.859104],
-              [-73.921754, 40.859094],
-              [-73.921713, 40.859096],
-              [-73.9217, 40.859084],
-              [-73.921671, 40.859067],
-              [-73.921656, 40.859076],
-              [-73.921641, 40.859068],
-              [-73.921623, 40.859056],
-              [-73.921576, 40.859055],
-              [-73.921563, 40.859038],
-              [-73.921537, 40.859028],
-              [-73.921517, 40.859012],
-              [-73.921496, 40.859006],
-              [-73.921415, 40.858952],
-              [-73.921407, 40.858939],
-              [-73.921395, 40.85894],
-              [-73.921362, 40.858919],
-              [-73.921323, 40.858907],
-              [-73.921307, 40.858899],
-              [-73.921296, 40.85888],
-              [-73.921314, 40.858875],
-              [-73.921345, 40.858885],
-              [-73.92137, 40.858897],
-              [-73.921388, 40.858897],
-              [-73.92137, 40.858879],
-              [-73.921333, 40.858867],
-              [-73.921309, 40.858864],
-              [-73.921278, 40.858869],
-              [-73.921256, 40.858847],
-              [-73.921239, 40.858815],
-              [-73.921239, 40.858794],
-              [-73.92125, 40.858784],
-              [-73.921264, 40.858737],
-              [-73.921257, 40.858715],
-              [-73.921239, 40.858694],
-              [-73.92121, 40.858689],
-              [-73.921173, 40.858646],
-              [-73.921133, 40.858639],
-              [-73.921104, 40.858616],
-              [-73.921109, 40.858598],
-              [-73.921121, 40.858588],
-              [-73.921103, 40.858572],
-              [-73.921089, 40.858567],
-              [-73.921086, 40.858556],
-              [-73.921072, 40.858553],
-              [-73.921102, 40.858513],
-              [-73.921122, 40.858445],
-              [-73.921092, 40.858441],
-              [-73.921068, 40.858417],
-              [-73.921059, 40.858375],
-              [-73.921079, 40.85833],
-              [-73.921059, 40.858317],
-              [-73.921079, 40.858312],
-              [-73.921088, 40.8583],
-              [-73.921064, 40.858294],
-              [-73.920997, 40.858304],
-              [-73.920987, 40.85829],
-              [-73.920942, 40.858284],
-              [-73.920915, 40.858219],
-              [-73.920938, 40.858121],
-              [-73.920812, 40.858082],
-              [-73.920664, 40.858036],
-              [-73.92066, 40.858027],
-              [-73.920624, 40.858017],
-              [-73.920609, 40.858019],
-              [-73.920484, 40.858233],
-              [-73.920351, 40.858195],
-              [-73.920365, 40.858165],
-              [-73.920473, 40.858195],
-              [-73.920483, 40.858172],
-              [-73.920425, 40.858147],
-              [-73.920431, 40.858137],
-              [-73.920494, 40.858165],
-              [-73.920508, 40.858142],
-              [-73.920508, 40.858135],
-              [-73.920474, 40.858126],
-              [-73.920479, 40.858112],
-              [-73.920519, 40.858126],
-              [-73.920557, 40.858058],
-              [-73.920524, 40.85805],
-              [-73.920531, 40.85804],
-              [-73.920559, 40.858049],
-              [-73.920583, 40.858011],
-              [-73.920544, 40.858001],
-              [-73.920566, 40.857964],
-              [-73.920531, 40.857953],
-              [-73.920536, 40.857945],
-              [-73.920568, 40.857956],
-              [-73.920599, 40.857899],
-              [-73.920581, 40.857884],
-              [-73.92057, 40.857889],
-              [-73.920565, 40.857883],
-              [-73.920583, 40.857874],
-              [-73.920612, 40.857896],
-              [-73.920612, 40.857904],
-              [-73.920773, 40.857941],
-              [-73.920776, 40.857903],
-              [-73.920724, 40.857885],
-              [-73.920751, 40.857832],
-              [-73.920758, 40.857811],
-              [-73.920714, 40.857795],
-              [-73.920718, 40.857783],
-              [-73.920706, 40.857778],
-              [-73.92068, 40.85782],
-              [-73.920673, 40.857817],
-              [-73.920695, 40.857781],
-              [-73.920631, 40.857796],
-              [-73.920629, 40.857785],
-              [-73.920694, 40.857768],
-              [-73.920694, 40.857757],
-              [-73.920679, 40.857753],
-              [-73.920674, 40.857744],
-              [-73.920683, 40.857728],
-              [-73.920706, 40.857701],
-              [-73.920715, 40.85771],
-              [-73.920726, 40.857712],
-              [-73.920741, 40.8577],
-              [-73.92074, 40.857712],
-              [-73.920712, 40.857756],
-              [-73.920708, 40.857767],
-              [-73.920718, 40.857772],
-              [-73.920721, 40.857767],
-              [-73.920746, 40.857775],
-              [-73.92075, 40.857763],
-              [-73.920859, 40.857796],
-              [-73.920855, 40.857808],
-              [-73.920825, 40.857801],
-              [-73.920819, 40.857813],
-              [-73.920909, 40.857844],
-              [-73.920918, 40.857817],
-              [-73.920959, 40.857827],
-              [-73.920955, 40.857835],
-              [-73.921008, 40.857848],
-              [-73.921014, 40.857839],
-              [-73.921002, 40.857818],
-              [-73.920983, 40.857802],
-              [-73.920988, 40.857776],
-              [-73.92096, 40.857787],
-              [-73.92095, 40.85781],
-              [-73.920824, 40.857752],
-              [-73.92084, 40.857721],
-              [-73.920963, 40.857779],
-              [-73.920985, 40.857764],
-              [-73.920989, 40.85775],
-              [-73.920976, 40.857736],
-              [-73.920963, 40.857742],
-              [-73.920929, 40.857727],
-              [-73.920931, 40.857723],
-              [-73.920921, 40.857716],
-              [-73.920917, 40.857721],
-              [-73.920911, 40.857719],
-              [-73.920915, 40.857714],
-              [-73.920905, 40.85771],
-              [-73.920902, 40.857714],
-              [-73.920892, 40.857713],
-              [-73.920894, 40.857707],
-              [-73.920875, 40.857701],
-              [-73.920872, 40.857707],
-              [-73.920864, 40.857706],
-              [-73.92087, 40.857694],
-              [-73.920897, 40.857702],
-              [-73.920904, 40.857694],
-              [-73.920911, 40.857695],
-              [-73.920907, 40.857703],
-              [-73.920917, 40.857707],
-              [-73.920924, 40.8577],
-              [-73.920932, 40.857703],
-              [-73.920926, 40.857711],
-              [-73.920936, 40.857715],
-              [-73.920959, 40.857685],
-              [-73.92089, 40.857664],
-              [-73.920914, 40.857619],
-              [-73.920892, 40.857613],
-              [-73.920885, 40.857624],
-              [-73.920814, 40.857607],
-              [-73.920788, 40.857654],
-              [-73.920773, 40.85765],
-              [-73.920798, 40.857604],
-              [-73.920776, 40.857598],
-              [-73.920789, 40.857572],
-              [-73.920784, 40.85757],
-              [-73.920719, 40.857633],
-              [-73.920697, 40.857621],
-              [-73.920755, 40.857557],
-              [-73.920772, 40.857549],
-              [-73.920789, 40.857562],
-              [-73.920794, 40.857553],
-              [-73.920901, 40.857582],
-              [-73.920894, 40.857598],
-              [-73.920919, 40.857606],
-              [-73.92093, 40.857592],
-              [-73.920982, 40.857612],
-              [-73.921002, 40.857585],
-              [-73.920946, 40.857564],
-              [-73.920951, 40.857557],
-              [-73.92101, 40.857578],
-              [-73.921015, 40.857574],
-              [-73.921011, 40.857571],
-              [-73.921017, 40.857562],
-              [-73.921025, 40.857564],
-              [-73.921032, 40.857555],
-              [-73.9209, 40.857494],
-              [-73.920909, 40.857479],
-              [-73.920937, 40.857487],
-              [-73.920945, 40.85747],
-              [-73.920997, 40.857484],
-              [-73.921036, 40.857439],
-              [-73.921014, 40.857429],
-              [-73.921049, 40.857382],
-              [-73.921069, 40.857394],
-              [-73.921094, 40.857383],
-              [-73.921164, 40.857437],
-              [-73.921132, 40.857455],
-              [-73.921073, 40.857517],
-              [-73.921121, 40.857537],
-              [-73.921184, 40.857524],
-              [-73.921223, 40.857522],
-              [-73.921305, 40.857421],
-              [-73.92152, 40.857349],
-              [-73.921584, 40.85729],
-              [-73.921558, 40.857153],
-              [-73.921559, 40.857089],
-              [-73.921592, 40.857065],
-              [-73.921703, 40.857021],
-              [-73.921837, 40.856993],
-              [-73.921863, 40.856992],
-              [-73.92197, 40.856911],
-              [-73.922045, 40.856854],
-              [-73.922075, 40.856779],
-              [-73.922157, 40.856766],
-              [-73.922169, 40.856738],
-              [-73.922146, 40.856726],
-              [-73.92211, 40.85672],
-              [-73.922167, 40.856663],
-              [-73.922247, 40.856614],
-              [-73.922293, 40.856627],
-              [-73.922377, 40.856584],
-              [-73.922345, 40.856253],
-              [-73.922376, 40.855978],
-              [-73.922345, 40.855682],
-              [-73.922345, 40.855682],
-              [-73.922345, 40.855682],
-              [-73.922179, 40.855897],
-              [-73.922181, 40.8559],
-              [-73.922183, 40.855903],
-              [-73.922184, 40.855907],
-              [-73.922184, 40.85591],
-              [-73.922184, 40.855914],
-              [-73.922183, 40.855917],
-              [-73.922181, 40.85592],
-              [-73.922179, 40.855923],
-              [-73.922176, 40.855926],
-              [-73.922172, 40.855928],
-              [-73.922168, 40.85593],
-              [-73.922164, 40.855931],
-              [-73.922159, 40.855932],
-              [-73.922155, 40.855932],
-              [-73.92215, 40.855932],
-              [-73.922063, 40.856054],
-              [-73.922077, 40.856059],
-              [-73.921873, 40.85635],
-              [-73.921888, 40.856355],
-              [-73.921818, 40.856454],
-              [-73.921835, 40.856462],
-              [-73.921765, 40.856566],
-              [-73.92177, 40.856568],
-              [-73.921698, 40.856672],
-              [-73.921519, 40.856601],
-              [-73.921591, 40.856497],
-              [-73.921597, 40.856499],
-              [-73.921672, 40.856394],
-              [-73.92169, 40.856401],
-              [-73.92176, 40.856305],
-              [-73.921823, 40.856329],
-              [-73.92203, 40.85604],
-              [-73.922042, 40.856045],
-              [-73.922129, 40.855923],
-              [-73.922127, 40.855919],
-              [-73.922126, 40.855916],
-              [-73.922125, 40.855912],
-              [-73.922125, 40.855909],
-              [-73.922126, 40.855905],
-              [-73.922127, 40.855902],
-              [-73.92213, 40.855899],
-              [-73.922133, 40.855896],
-              [-73.922136, 40.855893],
-              [-73.92214, 40.855891],
-              [-73.922144, 40.85589],
-              [-73.922149, 40.855889],
-              [-73.922153, 40.855889],
-              [-73.922158, 40.855889],
-              [-73.922236, 40.855771],
-              [-73.922296, 40.855664],
-              [-73.922279, 40.855655],
-              [-73.922298, 40.855599],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 168,
-        name: "Kips Bay",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.971682, 40.743041],
-              [-73.971889, 40.742725],
-              [-73.971741, 40.742664],
-              [-73.97196, 40.742338],
-              [-73.97229, 40.740721],
-              [-73.972196, 40.74071],
-              [-73.972183, 40.740706],
-              [-73.972168, 40.740693],
-              [-73.972165, 40.740676],
-              [-73.972196, 40.740403],
-              [-73.972217, 40.740308],
-              [-73.9723, 40.740035],
-              [-73.97231, 40.740019],
-              [-73.972331, 40.740011],
-              [-73.972439, 40.740022],
-              [-73.972657, 40.739223],
-              [-73.972657, 40.739221],
-              [-73.972657, 40.739221],
-              [-73.972568, 40.737404],
-              [-73.97249, 40.735803],
-              [-73.97251, 40.735787],
-              [-73.972869, 40.735771],
-              [-73.974414, 40.736414],
-              [-73.97446, 40.736294],
-              [-73.974417, 40.736278],
-              [-73.974412, 40.736285],
-              [-73.9744, 40.736284],
-              [-73.974027, 40.736126],
-              [-73.974047, 40.7361],
-              [-73.974429, 40.736261],
-              [-73.974428, 40.736266],
-              [-73.974466, 40.736278],
-              [-73.974504, 40.736065],
-              [-73.974504, 40.736065],
-              [-73.974475, 40.736062],
-              [-73.974441, 40.73605],
-              [-73.974434, 40.736054],
-              [-73.974141, 40.735956],
-              [-73.974151, 40.735934],
-              [-73.974447, 40.736033],
-              [-73.974448, 40.736038],
-              [-73.974478, 40.736044],
-              [-73.974505, 40.736042],
-              [-73.974515, 40.735895],
-              [-73.974515, 40.735895],
-              [-73.974515, 40.735893],
-              [-73.974508, 40.735797],
-              [-73.974487, 40.735877],
-              [-73.97445, 40.73587],
-              [-73.974461, 40.735832],
-              [-73.974249, 40.735798],
-              [-73.974245, 40.735781],
-              [-73.974258, 40.735767],
-              [-73.974469, 40.735799],
-              [-73.974469, 40.735789],
-              [-73.974477, 40.735789],
-              [-73.974465, 40.735668],
-              [-73.974391, 40.735669],
-              [-73.974391, 40.735673],
-              [-73.974263, 40.735673],
-              [-73.974264, 40.735677],
-              [-73.974192, 40.735681],
-              [-73.974189, 40.735661],
-              [-73.974157, 40.735659],
-              [-73.974145, 40.735673],
-              [-73.974092, 40.735642],
-              [-73.974089, 40.735646],
-              [-73.973419, 40.735368],
-              [-73.973439, 40.735341],
-              [-73.973057, 40.735177],
-              [-73.972807, 40.735524],
-              [-73.972937, 40.735599],
-              [-73.97293, 40.735608],
-              [-73.972763, 40.735505],
-              [-73.972997, 40.735187],
-              [-73.972964, 40.735169],
-              [-73.972932, 40.735205],
-              [-73.972885, 40.73518],
-              [-73.972864, 40.735212],
-              [-73.972837, 40.7352],
-              [-73.972945, 40.735065],
-              [-73.972967, 40.735078],
-              [-73.972945, 40.735108],
-              [-73.972996, 40.735128],
-              [-73.972971, 40.735157],
-              [-73.973003, 40.735173],
-              [-73.973037, 40.735128],
-              [-73.973299, 40.735242],
-              [-73.973299, 40.735241],
-              [-73.973322, 40.735202],
-              [-73.973299, 40.735177],
-              [-73.973279, 40.735148],
-              [-73.973272, 40.735119],
-              [-73.973277, 40.735079],
-              [-73.973296, 40.735041],
-              [-73.973329, 40.735011],
-              [-73.973369, 40.734989],
-              [-73.973418, 40.734971],
-              [-73.973459, 40.734968],
-              [-73.973523, 40.734975],
-              [-73.973545, 40.734944],
-              [-73.974437, 40.735316],
-              [-73.974437, 40.735315],
-              [-73.974437, 40.735315],
-              [-73.974378, 40.735081],
-              [-73.974378, 40.73508],
-              [-73.974648, 40.735082],
-              [-73.974907, 40.735313],
-              [-73.978527, 40.736854],
-              [-73.978527, 40.736855],
-              [-73.978527, 40.736856],
-              [-73.97852, 40.736858],
-              [-73.984742, 40.739503],
-              [-73.983381, 40.741375],
-              [-73.979711, 40.746385],
-              [-73.971682, 40.743041],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 172,
-        name: "Liberty Island",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-74.042554, 40.689963],
-              [-74.042639, 40.689928],
-              [-74.042694, 40.689997],
-              [-74.043468, 40.689637],
-              [-74.043516, 40.689191],
-              [-74.043641, 40.688767],
-              [-74.043975, 40.688582],
-              [-74.044385, 40.688516],
-              [-74.044784, 40.688596],
-              [-74.046275, 40.689327],
-              [-74.046364, 40.689433],
-              [-74.046803, 40.689953],
-              [-74.046831, 40.689939],
-              [-74.047477, 40.689611],
-              [-74.04773, 40.689915],
-              [-74.047586, 40.689983],
-              [-74.047431, 40.689804],
-              [-74.046918, 40.690047],
-              [-74.046892, 40.690059],
-              [-74.0472, 40.690425],
-              [-74.047111, 40.69047],
-              [-74.047116, 40.690656],
-              [-74.047183, 40.690747],
-              [-74.047195, 40.690763],
-              [-74.047213, 40.69079],
-              [-74.047226, 40.69082],
-              [-74.047232, 40.69085],
-              [-74.047232, 40.69088],
-              [-74.047225, 40.690911],
-              [-74.047213, 40.69094],
-              [-74.047194, 40.690967],
-              [-74.04717, 40.690991],
-              [-74.047141, 40.691011],
-              [-74.047108, 40.691026],
-              [-74.047073, 40.691038],
-              [-74.047035, 40.691046],
-              [-74.046997, 40.691049],
-              [-74.046172, 40.691098],
-              [-74.046147, 40.691123],
-              [-74.046096, 40.69112],
-              [-74.046044, 40.691114],
-              [-74.045994, 40.691105],
-              [-74.045926, 40.691083],
-              [-74.04586, 40.691056],
-              [-74.044606, 40.690573],
-              [-74.043908, 40.6902],
-              [-74.043878, 40.690188],
-              [-74.043506, 40.689687],
-              [-74.042735, 40.69005],
-              [-74.042784, 40.690121],
-              [-74.042704, 40.690155],
-              [-74.042554, 40.689963],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 174,
-        name: "Little Italy",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.994393, 40.719525],
-              [-73.994378, 40.719519],
-              [-73.994808, 40.718457],
-              [-73.996478, 40.719048],
-              [-73.997791, 40.716793],
-              [-73.998636, 40.717077],
-              [-73.999991, 40.717977],
-              [-73.99767, 40.72082],
-              [-73.994393, 40.719525],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 178,
-        name: "Lower East Side",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.974893, 40.715152],
-              [-73.97605, 40.712604],
-              [-73.976712, 40.711536],
-              [-73.977057, 40.711198],
-              [-73.977377, 40.710954],
-              [-73.977527, 40.710843],
-              [-73.977721, 40.71071],
-              [-73.977821, 40.710645],
-              [-73.977997, 40.710561],
-              [-73.978182, 40.710507],
-              [-73.979373, 40.710417],
-              [-73.980283, 40.710362],
-              [-73.980299, 40.710343],
-              [-73.980314, 40.710333],
-              [-73.980331, 40.710324],
-              [-73.98033, 40.710287],
-              [-73.98118, 40.710227],
-              [-73.981277, 40.710217],
-              [-73.981237, 40.709823],
-              [-73.981333, 40.709814],
-              [-73.988165, 40.709157],
-              [-73.988173, 40.709148],
-              [-73.98839, 40.709127],
-              [-73.988391, 40.709127],
-              [-73.988447, 40.709134],
-              [-73.98848, 40.709156],
-              [-73.988497, 40.709176],
-              [-73.988637, 40.709912],
-              [-73.988645, 40.709917],
-              [-73.98867, 40.70992],
-              [-73.988696, 40.709928],
-              [-73.988718, 40.709947],
-              [-73.988737, 40.709971],
-              [-73.988797, 40.709967],
-              [-73.989358, 40.709901],
-              [-73.990158, 40.70979],
-              [-73.990789, 40.70969],
-              [-73.991898, 40.709541],
-              [-73.992457, 40.711754],
-              [-73.992623, 40.713168],
-              [-73.99012, 40.713382],
-              [-73.9902, 40.714665],
-              [-73.989155, 40.716708],
-              [-73.994808, 40.718457],
-              [-73.994378, 40.719519],
-              [-73.994393, 40.719525],
-              [-73.992604, 40.724136],
-              [-73.978753, 40.719934],
-              [-73.977224, 40.71931],
-              [-73.975452, 40.718947],
-              [-73.974739, 40.718801],
-              [-73.973576, 40.718589],
-              [-73.974893, 40.715152],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 180,
-        name: "Marble Hill",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.908932, 40.872157],
-              [-73.908932, 40.872157],
-              [-73.909748, 40.873218],
-              [-73.910095, 40.873537],
-              [-73.910904, 40.874133],
-              [-73.910921, 40.874146],
-              [-73.912944, 40.874849],
-              [-73.913501, 40.875018],
-              [-73.913653, 40.875058],
-              [-73.914075, 40.875183],
-              [-73.914069, 40.875208],
-              [-73.915383, 40.875593],
-              [-73.915786, 40.875717],
-              [-73.914874, 40.876642],
-              [-73.914198, 40.876997],
-              [-73.912058, 40.878119],
-              [-73.911489, 40.879019],
-              [-73.910332, 40.879038],
-              [-73.909498, 40.878777],
-              [-73.909062, 40.878487],
-              [-73.908593, 40.878109],
-              [-73.90779, 40.877524],
-              [-73.907037, 40.876895],
-              [-73.906828, 40.876632],
-              [-73.906651, 40.875753],
-              [-73.907743, 40.872846],
-              [-73.908464, 40.872621],
-              [-73.90872, 40.872229],
-              [-73.908932, 40.872157],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 189,
-        name: "Midtown",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.962231, 40.754522],
-              [-73.967919, 40.747961],
-              [-73.979287, 40.752801],
-              [-73.982955, 40.747744],
-              [-73.979711, 40.746385],
-              [-73.983381, 40.741375],
-              [-73.991356, 40.744763],
-              [-73.987973, 40.749467],
-              [-73.987933, 40.749853],
-              [-73.993464, 40.75219],
-              [-73.990724, 40.75595],
-              [-73.985043, 40.753554],
-              [-73.979057, 40.761747],
-              [-73.984726, 40.764167],
-              [-73.982389, 40.767411],
-              [-73.982156, 40.767756],
-              [-73.982321, 40.767894],
-              [-73.982385, 40.768097],
-              [-73.982268, 40.768299],
-              [-73.982039, 40.768388],
-              [-73.981754, 40.7684],
-              [-73.981565, 40.768307],
-              [-73.981468, 40.76815],
-              [-73.981451, 40.767983],
-              [-73.98115, 40.767754],
-              [-73.973015, 40.764279],
-              [-73.958788, 40.758213],
-              [-73.959181, 40.757856],
-              [-73.960656, 40.755923],
-              [-73.962231, 40.754522],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 192,
-        name: "Morningside Heights",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.959647, 40.801156],
-              [-73.966701, 40.804169],
-              [-73.971037, 40.805918],
-              [-73.969331, 40.808107],
-              [-73.968463, 40.809319],
-              [-73.967402, 40.811022],
-              [-73.964325, 40.815513],
-              [-73.963485, 40.816584],
-              [-73.962054, 40.818095],
-              [-73.96179, 40.818025],
-              [-73.961038, 40.818704],
-              [-73.954288, 40.810992],
-              [-73.958181, 40.805565],
-              [-73.958249, 40.803107],
-              [-73.959642, 40.801163],
-              [-73.959647, 40.801156],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 199,
-        name: "Murray Hill",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.967919, 40.747961],
-              [-73.968174, 40.747667],
-              [-73.968365, 40.747433],
-              [-73.968232, 40.747368],
-              [-73.968296, 40.747296],
-              [-73.968302, 40.747273],
-              [-73.968316, 40.747274],
-              [-73.968326, 40.747252],
-              [-73.968943, 40.746518],
-              [-73.969624, 40.745747],
-              [-73.969697, 40.745657],
-              [-73.969876, 40.745445],
-              [-73.969906, 40.745458],
-              [-73.969912, 40.745452],
-              [-73.969931, 40.745463],
-              [-73.969927, 40.745468],
-              [-73.969974, 40.745496],
-              [-73.970048, 40.745403],
-              [-73.970339, 40.74505],
-              [-73.970326, 40.745037],
-              [-73.970326, 40.745023],
-              [-73.970378, 40.744952],
-              [-73.970443, 40.74487],
-              [-73.970461, 40.744865],
-              [-73.970472, 40.744866],
-              [-73.970788, 40.744406],
-              [-73.970941, 40.744471],
-              [-73.971125, 40.74419],
-              [-73.971111, 40.744184],
-              [-73.971092, 40.744182],
-              [-73.971075, 40.744176],
-              [-73.970905, 40.744103],
-              [-73.970758, 40.744039],
-              [-73.970751, 40.74403],
-              [-73.970718, 40.744017],
-              [-73.970706, 40.744003],
-              [-73.970703, 40.74399],
-              [-73.970743, 40.743948],
-              [-73.970751, 40.743949],
-              [-73.970761, 40.743942],
-              [-73.970792, 40.743956],
-              [-73.970799, 40.743953],
-              [-73.971173, 40.744108],
-              [-73.971358, 40.743838],
-              [-73.971205, 40.743772],
-              [-73.971281, 40.743652],
-              [-73.971241, 40.743638],
-              [-73.971208, 40.743699],
-              [-73.9712, 40.743706],
-              [-73.971188, 40.743707],
-              [-73.971098, 40.743678],
-              [-73.971085, 40.743664],
-              [-73.971233, 40.743415],
-              [-73.971246, 40.743402],
-              [-73.971256, 40.743402],
-              [-73.971267, 40.74339],
-              [-73.971354, 40.743425],
-              [-73.97135, 40.743432],
-              [-73.971357, 40.743437],
-              [-73.971361, 40.743444],
-              [-73.971311, 40.743524],
-              [-73.971356, 40.74354],
-              [-73.971681, 40.743042],
-              [-73.971682, 40.743041],
-              [-73.979711, 40.746385],
-              [-73.982955, 40.747744],
-              [-73.979287, 40.752801],
-              [-73.967919, 40.747961],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 206,
-        name: "NoHo",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.992604, 40.724136],
-              [-73.995395, 40.725004],
-              [-73.996772, 40.725432],
-              [-73.992854, 40.730083],
-              [-73.990757, 40.729741],
-              [-73.991285, 40.727762],
-              [-73.992604, 40.724136],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 207,
-        name: "Nolita",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.992604, 40.724136],
-              [-73.994393, 40.719525],
-              [-73.99767, 40.72082],
-              [-73.997124, 40.72229],
-              [-73.996661, 40.723467],
-              [-73.995395, 40.725004],
-              [-73.992604, 40.724136],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 247,
-        name: "Randall's Island",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.921338, 40.800852],
-              [-73.921033, 40.800265],
-              [-73.920315, 40.799375],
-              [-73.918629, 40.798631],
-              [-73.917766, 40.798309],
-              [-73.91663, 40.797858],
-              [-73.915455, 40.797063],
-              [-73.914438, 40.795666],
-              [-73.913804, 40.794512],
-              [-73.913791, 40.794087],
-              [-73.913784, 40.793836],
-              [-73.915141, 40.792012],
-              [-73.916306, 40.791117],
-              [-73.916171, 40.791014],
-              [-73.916865, 40.790362],
-              [-73.917045, 40.790442],
-              [-73.917106, 40.790469],
-              [-73.918563, 40.789695],
-              [-73.918509, 40.789425],
-              [-73.920685, 40.787028],
-              [-73.921889, 40.785701],
-              [-73.921992, 40.785754],
-              [-73.922188, 40.785509],
-              [-73.922528, 40.78451],
-              [-73.922852, 40.783556],
-              [-73.923065, 40.783343],
-              [-73.924011, 40.782661],
-              [-73.924121, 40.782523],
-              [-73.924236, 40.782497],
-              [-73.924481, 40.78232],
-              [-73.924673, 40.782245],
-              [-73.924884, 40.782223],
-              [-73.925061, 40.782216],
-              [-73.925242, 40.782274],
-              [-73.925389, 40.782274],
-              [-73.925718, 40.782177],
-              [-73.927324, 40.781091],
-              [-73.928101, 40.78092],
-              [-73.928493, 40.781008],
-              [-73.92862, 40.781084],
-              [-73.931054, 40.782397],
-              [-73.931301, 40.782435],
-              [-73.935023, 40.783006],
-              [-73.93526, 40.783217],
-              [-73.935813, 40.78371],
-              [-73.935904, 40.783889],
-              [-73.936071, 40.784217],
-              [-73.935922, 40.784826],
-              [-73.935869, 40.785041],
-              [-73.935713, 40.78568],
-              [-73.935528, 40.785895],
-              [-73.933456, 40.788299],
-              [-73.932707, 40.789167],
-              [-73.932407, 40.789719],
-              [-73.932345, 40.789709],
-              [-73.932243, 40.789693],
-              [-73.931476, 40.790479],
-              [-73.931136, 40.790753],
-              [-73.930842, 40.790991],
-              [-73.930729, 40.790955],
-              [-73.929724, 40.791366],
-              [-73.928719, 40.790944],
-              [-73.928595, 40.790967],
-              [-73.928073, 40.790823],
-              [-73.926923, 40.790789],
-              [-73.926418, 40.790693],
-              [-73.926254, 40.790695],
-              [-73.926104, 40.790732],
-              [-73.926062, 40.790828],
-              [-73.926066, 40.790937],
-              [-73.926086, 40.791004],
-              [-73.926089, 40.791092],
-              [-73.926045, 40.791177],
-              [-73.926048, 40.791265],
-              [-73.926113, 40.791413],
-              [-73.926154, 40.791548],
-              [-73.92616, 40.79162],
-              [-73.92616, 40.791716],
-              [-73.926273, 40.791758],
-              [-73.926443, 40.791797],
-              [-73.926548, 40.791828],
-              [-73.926569, 40.791859],
-              [-73.926514, 40.791898],
-              [-73.926279, 40.791879],
-              [-73.926146, 40.791848],
-              [-73.925928, 40.791876],
-              [-73.925765, 40.791858],
-              [-73.925636, 40.791823],
-              [-73.925588, 40.791822],
-              [-73.925581, 40.791856],
-              [-73.925608, 40.791897],
-              [-73.92617, 40.7919],
-              [-73.926429, 40.791957],
-              [-73.92663, 40.791947],
-              [-73.926739, 40.791929],
-              [-73.926702, 40.791833],
-              [-73.92664, 40.791791],
-              [-73.926579, 40.791794],
-              [-73.926508, 40.791773],
-              [-73.926368, 40.791662],
-              [-73.926331, 40.791504],
-              [-73.926345, 40.791248],
-              [-73.926488, 40.791022],
-              [-73.926566, 40.79098],
-              [-73.926703, 40.790968],
-              [-73.92702, 40.790999],
-              [-73.927115, 40.791015],
-              [-73.927166, 40.790997],
-              [-73.927241, 40.790997],
-              [-73.927316, 40.791007],
-              [-73.927381, 40.790997],
-              [-73.927439, 40.790961],
-              [-73.927653, 40.790958],
-              [-73.927784, 40.790942],
-              [-73.927869, 40.790955],
-              [-73.927893, 40.790979],
-              [-73.927911, 40.791039],
-              [-73.927894, 40.791091],
-              [-73.92783, 40.791161],
-              [-73.927459, 40.791592],
-              [-73.927321, 40.791743],
-              [-73.927233, 40.791791],
-              [-73.927136, 40.791929],
-              [-73.92765, 40.791982],
-              [-73.927801, 40.791997],
-              [-73.92826, 40.792323],
-              [-73.928267, 40.792329],
-              [-73.928338, 40.792857],
-              [-73.928342, 40.79298],
-              [-73.928323, 40.793102],
-              [-73.928265, 40.793213],
-              [-73.927353, 40.794462],
-              [-73.926936, 40.795497],
-              [-73.926903, 40.796032],
-              [-73.92689, 40.796252],
-              [-73.926952, 40.796708],
-              [-73.927118, 40.797204],
-              [-73.927185, 40.79735],
-              [-73.927355, 40.797317],
-              [-73.927347, 40.797294],
-              [-73.927671, 40.797217],
-              [-73.927707, 40.797299],
-              [-73.927383, 40.797377],
-              [-73.927367, 40.797347],
-              [-73.927203, 40.797387],
-              [-73.927217, 40.797415],
-              [-73.927277, 40.797407],
-              [-73.927439, 40.797711],
-              [-73.927476, 40.798123],
-              [-73.927263, 40.798919],
-              [-73.926964, 40.79972],
-              [-73.926883, 40.799941],
-              [-73.926854, 40.799988],
-              [-73.926696, 40.800147],
-              [-73.926736, 40.800165],
-              [-73.927248, 40.800393],
-              [-73.927115, 40.800564],
-              [-73.927067, 40.800543],
-              [-73.927169, 40.800406],
-              [-73.926707, 40.800201],
-              [-73.926666, 40.800183],
-              [-73.926554, 40.800315],
-              [-73.926743, 40.800403],
-              [-73.926737, 40.800411],
-              [-73.926725, 40.800407],
-              [-73.926721, 40.800414],
-              [-73.926538, 40.800334],
-              [-73.926503, 40.800375],
-              [-73.926685, 40.800455],
-              [-73.926552, 40.800605],
-              [-73.926845, 40.80074],
-              [-73.926832, 40.800766],
-              [-73.926869, 40.800784],
-              [-73.926995, 40.800636],
-              [-73.927046, 40.800652],
-              [-73.926886, 40.800841],
-              [-73.92651, 40.800657],
-              [-73.925776, 40.801731],
-              [-73.925698, 40.801811],
-              [-73.925493, 40.801956],
-              [-73.925275, 40.801999],
-              [-73.92436, 40.801986],
-              [-73.923784, 40.801947],
-              [-73.922634, 40.801869],
-              [-73.921911, 40.801503],
-              [-73.921338, 40.800852],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 257,
-        name: "Roosevelt Island",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.9449, 40.769484],
-              [-73.944724, 40.769734],
-              [-73.944725, 40.769786],
-              [-73.944562, 40.769886],
-              [-73.94427, 40.770141],
-              [-73.944004, 40.770294],
-              [-73.943204, 40.770992],
-              [-73.942578, 40.771702],
-              [-73.942127, 40.772035],
-              [-73.94137, 40.772309],
-              [-73.940887, 40.772373],
-              [-73.94057, 40.772615],
-              [-73.940457, 40.772741],
-              [-73.940275, 40.772947],
-              [-73.94018, 40.772959],
-              [-73.940077, 40.772926],
-              [-73.940012, 40.772842],
-              [-73.940086, 40.772722],
-              [-73.940185, 40.772565],
-              [-73.940282, 40.772301],
-              [-73.94028, 40.772079],
-              [-73.940279, 40.77183],
-              [-73.9402, 40.771128],
-              [-73.940214, 40.770996],
-              [-73.940229, 40.770852],
-              [-73.94028, 40.770671],
-              [-73.940355, 40.770497],
-              [-73.940826, 40.769912],
-              [-73.941118, 40.769414],
-              [-73.941219, 40.769334],
-              [-73.941533, 40.769177],
-              [-73.9418, 40.769047],
-              [-73.942855, 40.768303],
-              [-73.94484, 40.765824],
-              [-73.949432, 40.760549],
-              [-73.95037, 40.759816],
-              [-73.95126, 40.75881],
-              [-73.951581, 40.758581],
-              [-73.951915, 40.758225],
-              [-73.952177, 40.757831],
-              [-73.952211, 40.757584],
-              [-73.95234, 40.757393],
-              [-73.952541, 40.757095],
-              [-73.952903, 40.756739],
-              [-73.955073, 40.754799],
-              [-73.955195, 40.754775],
-              [-73.956924, 40.752961],
-              [-73.957915, 40.75211],
-              [-73.958608, 40.751613],
-              [-73.95904, 40.751392],
-              [-73.959028, 40.751297],
-              [-73.959643, 40.750752],
-              [-73.96004, 40.750473],
-              [-73.960392, 40.750256],
-              [-73.960829, 40.749911],
-              [-73.961174, 40.749498],
-              [-73.961425, 40.749401],
-              [-73.961583, 40.749414],
-              [-73.961608, 40.749564],
-              [-73.961551, 40.749774],
-              [-73.96132, 40.750075],
-              [-73.961111, 40.7506],
-              [-73.9608, 40.751025],
-              [-73.960799, 40.751026],
-              [-73.960709, 40.751148],
-              [-73.960546, 40.751308],
-              [-73.960011, 40.751832],
-              [-73.959685, 40.752552],
-              [-73.959501, 40.752956],
-              [-73.959016, 40.753595],
-              [-73.958927, 40.753562],
-              [-73.958477, 40.753937],
-              [-73.958458, 40.753949],
-              [-73.958183, 40.754267],
-              [-73.955348, 40.757538],
-              [-73.955274, 40.757624],
-              [-73.953662, 40.75948],
-              [-73.953323, 40.759746],
-              [-73.953391, 40.759797],
-              [-73.953125, 40.760053],
-              [-73.952978, 40.759962],
-              [-73.952086, 40.760862],
-              [-73.951625, 40.761679],
-              [-73.950937, 40.76257],
-              [-73.949972, 40.763646],
-              [-73.949204, 40.76427],
-              [-73.949086, 40.764264],
-              [-73.947964, 40.765688],
-              [-73.945915, 40.768214],
-              [-73.945851, 40.768248],
-              [-73.945845, 40.768321],
-              [-73.945754, 40.768302],
-              [-73.945704, 40.768328],
-              [-73.945504, 40.768594],
-              [-73.945481, 40.768625],
-              [-73.945627, 40.768693],
-              [-73.945563, 40.768802],
-              [-73.945405, 40.768732],
-              [-73.945382, 40.768763],
-              [-73.945245, 40.768773],
-              [-73.945157, 40.769118],
-              [-73.9449, 40.769484],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 266,
-        name: "SoHo",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-74.002856, 40.728347],
-              [-73.997635, 40.725786],
-              [-73.996772, 40.725432],
-              [-73.995395, 40.725004],
-              [-73.996661, 40.723467],
-              [-73.997124, 40.72229],
-              [-73.99767, 40.72082],
-              [-73.999991, 40.717977],
-              [-74.001886, 40.719395],
-              [-74.008302, 40.723977],
-              [-74.01106, 40.725744],
-              [-74.011376, 40.72428],
-              [-74.011965, 40.724332],
-              [-74.011963, 40.724342],
-              [-74.011653, 40.725872],
-              [-74.01166, 40.725873],
-              [-74.011958, 40.725978],
-              [-74.012316, 40.726034],
-              [-74.015163, 40.72632],
-              [-74.015204, 40.726362],
-              [-74.015124, 40.726794],
-              [-74.011542, 40.726447],
-              [-74.011383, 40.728229],
-              [-74.010799, 40.728196],
-              [-74.010659, 40.729832],
-              [-74.009208, 40.729706],
-              [-74.00935, 40.728988],
-              [-74.002856, 40.728347],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 277,
-        name: "Stuyvesant Town",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.974378, 40.735081],
-              [-73.974378, 40.735081],
-              [-73.973959, 40.733421],
-              [-73.973943, 40.733403],
-              [-73.973907, 40.73335],
-              [-73.973866, 40.733311],
-              [-73.973845, 40.733265],
-              [-73.973865, 40.733246],
-              [-73.973857, 40.733216],
-              [-73.973841, 40.73322],
-              [-73.973849, 40.733233],
-              [-73.973831, 40.733246],
-              [-73.97381, 40.733249],
-              [-73.973784, 40.733227],
-              [-73.973803, 40.733212],
-              [-73.973784, 40.7332],
-              [-73.973765, 40.733213],
-              [-73.973727, 40.73322],
-              [-73.97369, 40.733219],
-              [-73.973676, 40.733206],
-              [-73.973692, 40.733174],
-              [-73.973737, 40.733143],
-              [-73.973819, 40.733099],
-              [-73.973872, 40.733095],
-              [-73.973868, 40.733066],
-              [-73.973659, 40.732245],
-              [-73.973515, 40.73168],
-              [-73.973431, 40.731351],
-              [-73.973413, 40.731282],
-              [-73.971963, 40.73],
-              [-73.971956, 40.729988],
-              [-73.97148, 40.72925],
-              [-73.971427, 40.728486],
-              [-73.971556, 40.727703],
-              [-73.971569, 40.727694],
-              [-73.971573, 40.727646],
-              [-73.97159, 40.72764],
-              [-73.971619, 40.727651],
-              [-73.971652, 40.727643],
-              [-73.971685, 40.727406],
-              [-73.971537, 40.727393],
-              [-73.971521, 40.727386],
-              [-73.971513, 40.727374],
-              [-73.971493, 40.727373],
-              [-73.97149, 40.72735],
-              [-73.971629, 40.726761],
-              [-73.982553, 40.731375],
-              [-73.982022, 40.732012],
-              [-73.978527, 40.736855],
-              [-73.978527, 40.736854],
-              [-73.974907, 40.735313],
-              [-73.974648, 40.735082],
-              [-73.974378, 40.73508],
-              [-73.974378, 40.735081],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 280,
-        name: "Theater District",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.985043, 40.753554],
-              [-73.990724, 40.75595],
-              [-73.984726, 40.764167],
-              [-73.979057, 40.761747],
-              [-73.985043, 40.753554],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 286,
-        name: "Tribeca",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-74.001886, 40.719395],
-              [-74.003533, 40.717449],
-              [-74.008614, 40.711372],
-              [-74.013754, 40.71369],
-              [-74.01244, 40.719058],
-              [-74.013036, 40.719086],
-              [-74.013031, 40.719131],
-              [-74.012961, 40.719484],
-              [-74.012914, 40.719708],
-              [-74.012945, 40.719713],
-              [-74.012946, 40.719697],
-              [-74.013026, 40.719709],
-              [-74.013024, 40.719726],
-              [-74.013445, 40.719782],
-              [-74.013447, 40.719782],
-              [-74.013399, 40.720032],
-              [-74.01639, 40.720379],
-              [-74.016315, 40.720739],
-              [-74.013963, 40.720464],
-              [-74.012966, 40.720329],
-              [-74.012964, 40.720335],
-              [-74.012869, 40.720834],
-              [-74.013239, 40.720888],
-              [-74.015695, 40.72122],
-              [-74.015626, 40.721551],
-              [-74.013138, 40.721251],
-              [-74.01308, 40.721546],
-              [-74.013044, 40.721545],
-              [-74.01301, 40.72156],
-              [-74.012544, 40.721504],
-              [-74.01254, 40.721503],
-              [-74.012539, 40.721503],
-              [-74.012504, 40.721672],
-              [-74.012492, 40.72178],
-              [-74.011965, 40.724332],
-              [-74.011376, 40.72428],
-              [-74.01106, 40.725744],
-              [-74.008302, 40.723977],
-              [-74.001886, 40.719395],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 287,
-        name: "Two Bridges",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.991898, 40.709541],
-              [-73.992996, 40.709359],
-              [-73.99374, 40.709246],
-              [-73.994141, 40.709233],
-              [-73.994378, 40.709193],
-              [-73.994579, 40.709131],
-              [-73.995377, 40.709006],
-              [-73.997283, 40.708732],
-              [-73.998084, 40.70851],
-              [-73.99909, 40.707975],
-              [-73.999102, 40.707968],
-              [-73.999102, 40.707968],
-              [-74.003337, 40.711236],
-              [-74.002363, 40.711591],
-              [-73.99777, 40.711971],
-              [-73.997994, 40.712722],
-              [-73.992623, 40.713168],
-              [-73.992457, 40.711754],
-              [-73.991898, 40.709541],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 290,
-        name: "Upper East Side",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.943592, 40.782748],
-              [-73.943648, 40.782656],
-              [-73.943871, 40.781273],
-              [-73.943459, 40.780048],
-              [-73.943214, 40.779318],
-              [-73.943004, 40.779639],
-              [-73.942716, 40.779544],
-              [-73.942712, 40.779215],
-              [-73.942536, 40.779091],
-              [-73.942893, 40.778614],
-              [-73.942438, 40.777315],
-              [-73.942245, 40.777104],
-              [-73.942074, 40.776918],
-              [-73.942003, 40.776185],
-              [-73.94262, 40.775181],
-              [-73.942856, 40.774797],
-              [-73.94293, 40.774676],
-              [-73.945871, 40.771692],
-              [-73.946619, 40.770933],
-              [-73.948664, 40.768858],
-              [-73.95007, 40.767025],
-              [-73.954418, 40.762184],
-              [-73.956508, 40.760285],
-              [-73.958788, 40.758213],
-              [-73.973015, 40.764279],
-              [-73.95576, 40.787907],
-              [-73.944023, 40.78296],
-              [-73.943592, 40.782748],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 291,
-        name: "Upper West Side",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.958358, 40.800369],
-              [-73.981439, 40.768763],
-              [-73.981549, 40.768559],
-              [-73.981565, 40.768307],
-              [-73.981754, 40.7684],
-              [-73.982039, 40.768388],
-              [-73.982268, 40.768299],
-              [-73.982385, 40.768097],
-              [-73.982321, 40.767894],
-              [-73.982156, 40.767756],
-              [-73.982389, 40.767411],
-              [-73.99365, 40.772146],
-              [-73.994159, 40.772493],
-              [-73.993831, 40.772932],
-              [-73.993891, 40.772955],
-              [-73.993963, 40.772945],
-              [-73.994013, 40.772883],
-              [-73.994122, 40.772924],
-              [-73.994137, 40.772902],
-              [-73.994301, 40.77297],
-              [-73.994282, 40.772994],
-              [-73.994377, 40.77304],
-              [-73.994294, 40.773156],
-              [-73.995023, 40.773481],
-              [-73.995089, 40.773388],
-              [-73.995014, 40.773358],
-              [-73.99505, 40.773297],
-              [-73.996241, 40.77379],
-              [-73.996196, 40.773852],
-              [-73.996099, 40.773952],
-              [-73.996179, 40.773987],
-              [-73.996095, 40.774086],
-              [-73.995572, 40.773871],
-              [-73.994017, 40.773214],
-              [-73.993936, 40.77318],
-              [-73.993862, 40.77327],
-              [-73.993822, 40.773382],
-              [-73.993767, 40.773484],
-              [-73.993698, 40.773562],
-              [-73.993358, 40.773927],
-              [-73.992623, 40.774974],
-              [-73.992578, 40.774956],
-              [-73.992528, 40.775002],
-              [-73.99247, 40.775024],
-              [-73.992404, 40.775018],
-              [-73.992267, 40.775116],
-              [-73.992178, 40.775279],
-              [-73.992059, 40.775498],
-              [-73.992125, 40.775509],
-              [-73.992227, 40.775482],
-              [-73.992329, 40.775469],
-              [-73.992362, 40.775502],
-              [-73.992386, 40.775557],
-              [-73.992088, 40.775984],
-              [-73.990927, 40.777567],
-              [-73.990396, 40.777585],
-              [-73.989461, 40.778875],
-              [-73.989176, 40.779288],
-              [-73.988869, 40.779693],
-              [-73.988872, 40.779714],
-              [-73.989219, 40.779698],
-              [-73.989278, 40.779723],
-              [-73.989409, 40.779738],
-              [-73.989499, 40.779725],
-              [-73.989596, 40.779698],
-              [-73.98968, 40.779678],
-              [-73.989753, 40.779671],
-              [-73.989842, 40.779681],
-              [-73.99004, 40.779708],
-              [-73.990138, 40.7797],
-              [-73.990336, 40.779662],
-              [-73.990431, 40.779665],
-              [-73.990622, 40.779676],
-              [-73.990745, 40.779671],
-              [-73.990872, 40.779646],
-              [-73.990962, 40.77964],
-              [-73.991057, 40.779652],
-              [-73.991157, 40.77967],
-              [-73.991243, 40.779671],
-              [-73.991255, 40.779651],
-              [-73.991295, 40.77963],
-              [-73.991322, 40.779632],
-              [-73.991359, 40.779586],
-              [-73.991551, 40.779575],
-              [-73.99142, 40.779755],
-              [-73.988886, 40.779879],
-              [-73.98894, 40.779956],
-              [-73.988926, 40.780059],
-              [-73.988912, 40.780096],
-              [-73.988919, 40.780226],
-              [-73.988381, 40.780981],
-              [-73.988232, 40.781233],
-              [-73.98821, 40.781225],
-              [-73.98814, 40.781409],
-              [-73.988041, 40.781586],
-              [-73.9881, 40.781603],
-              [-73.988076, 40.781651],
-              [-73.988018, 40.781634],
-              [-73.987961, 40.781771],
-              [-73.985466, 40.785361],
-              [-73.986173, 40.786068],
-              [-73.986456, 40.785919],
-              [-73.987072, 40.78519],
-              [-73.987119, 40.78521],
-              [-73.986498, 40.785951],
-              [-73.986165, 40.786122],
-              [-73.986128, 40.786239],
-              [-73.986071, 40.786241],
-              [-73.986027, 40.786229],
-              [-73.986098, 40.786058],
-              [-73.985429, 40.785414],
-              [-73.985081, 40.785922],
-              [-73.985199, 40.785967],
-              [-73.985171, 40.786013],
-              [-73.985216, 40.786031],
-              [-73.985255, 40.785976],
-              [-73.985243, 40.785973],
-              [-73.98525, 40.785963],
-              [-73.985282, 40.785979],
-              [-73.98524, 40.786036],
-              [-73.985684, 40.786222],
-              [-73.985718, 40.786176],
-              [-73.985766, 40.786196],
-              [-73.985683, 40.78631],
-              [-73.985636, 40.78629],
-              [-73.985671, 40.786243],
-              [-73.985205, 40.786048],
-              [-73.985211, 40.78604],
-              [-73.985163, 40.786021],
-              [-73.985132, 40.78606],
-              [-73.985017, 40.786014],
-              [-73.984655, 40.786535],
-              [-73.985744, 40.78657],
-              [-73.985892, 40.786427],
-              [-73.985943, 40.786453],
-              [-73.98595, 40.786487],
-              [-73.985812, 40.786617],
-              [-73.985135, 40.786588],
-              [-73.984619, 40.786586],
-              [-73.981952, 40.790394],
-              [-73.972823, 40.803428],
-              [-73.971037, 40.805918],
-              [-73.966701, 40.804169],
-              [-73.959647, 40.801156],
-              [-73.958509, 40.800682],
-              [-73.958533, 40.800491],
-              [-73.958358, 40.800369],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 296,
-        name: "Washington Heights",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-73.922298, 40.855599],
-              [-73.924069, 40.853319],
-              [-73.924441, 40.85284],
-              [-73.927787, 40.848017],
-              [-73.929719, 40.845232],
-              [-73.930543, 40.843344],
-              [-73.930895, 40.842538],
-              [-73.934915, 40.835231],
-              [-73.935157, 40.832682],
-              [-73.934986, 40.831369],
-              [-73.934871, 40.830482],
-              [-73.934568, 40.828149],
-              [-73.940346, 40.830458],
-              [-73.948905, 40.834066],
-              [-73.949342, 40.833982],
-              [-73.95022, 40.834305],
-              [-73.950155, 40.834397],
-              [-73.94998, 40.834876],
-              [-73.949889, 40.835285],
-              [-73.949743, 40.835628],
-              [-73.949739, 40.835706],
-              [-73.949712, 40.835775],
-              [-73.949693, 40.835863],
-              [-73.949683, 40.835922],
-              [-73.949691, 40.836074],
-              [-73.949686, 40.83611],
-              [-73.949645, 40.83622],
-              [-73.949621, 40.836232],
-              [-73.949601, 40.836305],
-              [-73.949583, 40.836484],
-              [-73.949546, 40.836604],
-              [-73.94951, 40.836665],
-              [-73.949457, 40.836706],
-              [-73.949389, 40.836716],
-              [-73.949328, 40.836719],
-              [-73.94929, 40.836729],
-              [-73.949283, 40.836741],
-              [-73.949284, 40.836754],
-              [-73.949293, 40.836764],
-              [-73.949303, 40.836769],
-              [-73.949377, 40.83678],
-              [-73.949442, 40.836792],
-              [-73.949463, 40.836798],
-              [-73.949482, 40.836806],
-              [-73.949497, 40.836815],
-              [-73.949511, 40.836826],
-              [-73.949522, 40.836838],
-              [-73.949531, 40.836851],
-              [-73.949537, 40.836865],
-              [-73.94954, 40.83688],
-              [-73.949513, 40.836952],
-              [-73.949181, 40.837864],
-              [-73.949152, 40.837934],
-              [-73.949005, 40.838426],
-              [-73.948929, 40.838643],
-              [-73.94884, 40.838876],
-              [-73.948758, 40.83902],
-              [-73.948561, 40.839565],
-              [-73.948496, 40.839668],
-              [-73.948429, 40.839861],
-              [-73.948334, 40.839997],
-              [-73.948234, 40.840118],
-              [-73.948075, 40.840466],
-              [-73.947921, 40.840758],
-              [-73.947709, 40.841264],
-              [-73.947606, 40.84141],
-              [-73.947383, 40.841941],
-              [-73.947274, 40.842052],
-              [-73.94721, 40.842101],
-              [-73.947168, 40.842159],
-              [-73.947134, 40.842259],
-              [-73.947121, 40.842286],
-              [-73.947107, 40.842307],
-              [-73.94696, 40.84254],
-              [-73.946935, 40.842619],
-              [-73.946788, 40.842815],
-              [-73.946755, 40.842888],
-              [-73.94671, 40.842944],
-              [-73.946616, 40.84302],
-              [-73.946459, 40.843254],
-              [-73.946369, 40.843455],
-              [-73.946124, 40.843892],
-              [-73.946119, 40.844005],
-              [-73.946097, 40.844055],
-              [-73.946094, 40.844097],
-              [-73.946114, 40.844196],
-              [-73.946117, 40.844225],
-              [-73.946093, 40.844291],
-              [-73.946105, 40.84435],
-              [-73.946084, 40.844468],
-              [-73.946091, 40.844499],
-              [-73.946079, 40.844598],
-              [-73.946109, 40.844859],
-              [-73.946123, 40.84488],
-              [-73.946128, 40.844912],
-              [-73.946147, 40.84494],
-              [-73.94616, 40.84503],
-              [-73.946161, 40.845126],
-              [-73.946223, 40.845259],
-              [-73.946211, 40.845334],
-              [-73.946172, 40.845393],
-              [-73.946207, 40.845462],
-              [-73.946231, 40.845577],
-              [-73.946277, 40.845691],
-              [-73.946308, 40.845741],
-              [-73.946439, 40.845886],
-              [-73.946466, 40.845881],
-              [-73.94648, 40.84592],
-              [-73.946445, 40.845921],
-              [-73.946517, 40.845995],
-              [-73.946504, 40.846028],
-              [-73.946456, 40.846058],
-              [-73.946426, 40.846063],
-              [-73.946348, 40.846156],
-              [-73.946351, 40.84624],
-              [-73.946474, 40.846529],
-              [-73.946517, 40.846593],
-              [-73.946641, 40.846685],
-              [-73.946718, 40.846825],
-              [-73.94679, 40.846997],
-              [-73.946824, 40.847134],
-              [-73.946809, 40.847288],
-              [-73.946782, 40.847353],
-              [-73.946699, 40.847497],
-              [-73.946686, 40.847728],
-              [-73.946656, 40.847883],
-              [-73.946623, 40.848146],
-              [-73.946619, 40.848278],
-              [-73.946644, 40.848399],
-              [-73.946651, 40.8485],
-              [-73.94665, 40.848668],
-              [-73.946679, 40.848837],
-              [-73.946636, 40.848912],
-              [-73.946839, 40.849383],
-              [-73.946855, 40.849428],
-              [-73.946967, 40.849815],
-              [-73.947013, 40.85],
-              [-73.947028, 40.850042],
-              [-73.94702, 40.850094],
-              [-73.947044, 40.850255],
-              [-73.946998, 40.850398],
-              [-73.946964, 40.850466],
-              [-73.946926, 40.850528],
-              [-73.946516, 40.850968],
-              [-73.946502, 40.850976],
-              [-73.946472, 40.850985],
-              [-73.946442, 40.850988],
-              [-73.946412, 40.850985],
-              [-73.946372, 40.85097],
-              [-73.946339, 40.850948],
-              [-73.946321, 40.850923],
-              [-73.946329, 40.850904],
-              [-73.946352, 40.850893],
-              [-73.946352, 40.85087],
-              [-73.946338, 40.850853],
-              [-73.946327, 40.850846],
-              [-73.946316, 40.850839],
-              [-73.946297, 40.850833],
-              [-73.946275, 40.850841],
-              [-73.946244, 40.850885],
-              [-73.946198, 40.850906],
-              [-73.946173, 40.850958],
-              [-73.946196, 40.850974],
-              [-73.946201, 40.851031],
-              [-73.946186, 40.851031],
-              [-73.946172, 40.851052],
-              [-73.946158, 40.851047],
-              [-73.946147, 40.851031],
-              [-73.946145, 40.851007],
-              [-73.946134, 40.850984],
-              [-73.94612, 40.850969],
-              [-73.946097, 40.850966],
-              [-73.946082, 40.850981],
-              [-73.946087, 40.851008],
-              [-73.946097, 40.851033],
-              [-73.946142, 40.851094],
-              [-73.946137, 40.851155],
-              [-73.946117, 40.851173],
-              [-73.946048, 40.851189],
-              [-73.945965, 40.851172],
-              [-73.94593, 40.85119],
-              [-73.945896, 40.851187],
-              [-73.945841, 40.851169],
-              [-73.945767, 40.851169],
-              [-73.945725, 40.851197],
-              [-73.945728, 40.851228],
-              [-73.945683, 40.851227],
-              [-73.945653, 40.851211],
-              [-73.945612, 40.85124],
-              [-73.945556, 40.851321],
-              [-73.94552, 40.851362],
-              [-73.945449, 40.851428],
-              [-73.945389, 40.851453],
-              [-73.945266, 40.851571],
-              [-73.944978, 40.851732],
-              [-73.944667, 40.851921],
-              [-73.944577, 40.851943],
-              [-73.944504, 40.851957],
-              [-73.944397, 40.851893],
-              [-73.944325, 40.851823],
-              [-73.944317, 40.851856],
-              [-73.944197, 40.852015],
-              [-73.943882, 40.852172],
-              [-73.94348, 40.852282],
-              [-73.94187, 40.853868],
-              [-73.940293, 40.856642],
-              [-73.939585, 40.85748],
-              [-73.939111, 40.858041],
-              [-73.938814, 40.858537],
-              [-73.93831, 40.859381],
-              [-73.937815, 40.859834],
-              [-73.937789, 40.859812],
-              [-73.935532, 40.858918],
-              [-73.933053, 40.859828],
-              [-73.931313, 40.859328],
-              [-73.930122, 40.858744],
-              [-73.927625, 40.859012],
-              [-73.927091, 40.85976],
-              [-73.92471, 40.861552],
-              [-73.92293, 40.858941],
-              [-73.922906, 40.855912],
-              [-73.922531, 40.85568],
-              [-73.922298, 40.855599],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        cartodb_id: 299,
-        name: "West Village",
-        created_at: "2014-03-09T19:11:53Z",
-        updated_at: "2014-03-09T19:11:53Z",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-74.00881, 40.742378],
-              [-73.996797, 40.737364],
-              [-74.001413, 40.731065],
-              [-74.002856, 40.728347],
-              [-74.00935, 40.728988],
-              [-74.009208, 40.729706],
-              [-74.010659, 40.729832],
-              [-74.010799, 40.728196],
-              [-74.011383, 40.728229],
-              [-74.014391, 40.728463],
-              [-74.014345, 40.728622],
-              [-74.014077, 40.73064],
-              [-74.014087, 40.730663],
-              [-74.014038, 40.730694],
-              [-74.014016, 40.730683],
-              [-74.011163, 40.730451],
-              [-74.011162, 40.730451],
-              [-74.011068, 40.731545],
-              [-74.010984, 40.732522],
-              [-74.010941, 40.73302],
-              [-74.01166, 40.733032],
-              [-74.014002, 40.733068],
-              [-74.014, 40.733332],
-              [-74.014, 40.733332],
-              [-74.010904, 40.733298],
-              [-74.010903, 40.733298],
-              [-74.01087, 40.733593],
-              [-74.010833, 40.734007],
-              [-74.012018, 40.734064],
-              [-74.012001, 40.734338],
-              [-74.012001, 40.734338],
-              [-74.010827, 40.734284],
-              [-74.010814, 40.734283],
-              [-74.010643, 40.736049],
-              [-74.010677, 40.736056],
-              [-74.01067, 40.736102],
-              [-74.010688, 40.736114],
-              [-74.010704, 40.736129],
-              [-74.010717, 40.736146],
-              [-74.010726, 40.736165],
-              [-74.010731, 40.736184],
-              [-74.010732, 40.736204],
-              [-74.010728, 40.736224],
-              [-74.01072, 40.736243],
-              [-74.010709, 40.73626],
-              [-74.010694, 40.736276],
-              [-74.010676, 40.73629],
-              [-74.010655, 40.736301],
-              [-74.010654, 40.736348],
-              [-74.010619, 40.736345],
-              [-74.010477, 40.738064],
-              [-74.011087, 40.738093],
-              [-74.011068, 40.738304],
-              [-74.011089, 40.738318],
-              [-74.011107, 40.738333],
-              [-74.011121, 40.738351],
-              [-74.011132, 40.73837],
-              [-74.011139, 40.73839],
-              [-74.011116, 40.738392],
-              [-74.011092, 40.738392],
-              [-74.011054, 40.738388],
-              [-74.011051, 40.738396],
-              [-74.010454, 40.738365],
-              [-74.010393, 40.739174],
-              [-74.010499, 40.739173],
-              [-74.010555, 40.739186],
-              [-74.010747, 40.739176],
-              [-74.011388, 40.739216],
-              [-74.011485, 40.73926],
-              [-74.01158, 40.739265],
-              [-74.011588, 40.739272],
-              [-74.011587, 40.739479],
-              [-74.011634, 40.739506],
-              [-74.011628, 40.739562],
-              [-74.011555, 40.739557],
-              [-74.011527, 40.739874],
-              [-74.011593, 40.739874],
-              [-74.011613, 40.739708],
-              [-74.011616, 40.739708],
-              [-74.012348, 40.739738],
-              [-74.012369, 40.739752],
-              [-74.012374, 40.739779],
-              [-74.012349, 40.739794],
-              [-74.012286, 40.739793],
-              [-74.012289, 40.739899],
-              [-74.012314, 40.739903],
-              [-74.012343, 40.739921],
-              [-74.012357, 40.739937],
-              [-74.012349, 40.740016],
-              [-74.012324, 40.740031],
-              [-74.01227, 40.740034],
-              [-74.012263, 40.740142],
-              [-74.012309, 40.740148],
-              [-74.012329, 40.740163],
-              [-74.01233, 40.740183],
-              [-74.01231, 40.740195],
-              [-74.011577, 40.74016],
-              [-74.011586, 40.739974],
-              [-74.011515, 40.739973],
-              [-74.011461, 40.740584],
-              [-74.012157, 40.740606],
-              [-74.012154, 40.740674],
-              [-74.012657, 40.740697],
-              [-74.012659, 40.740748],
-              [-74.011475, 40.740687],
-              [-74.011374, 40.740724],
-              [-74.009631, 40.740645],
-              [-74.009497, 40.740727],
-              [-74.009492, 40.740743],
-              [-74.009516, 40.740774],
-              [-74.009518, 40.740806],
-              [-74.009463, 40.740902],
-              [-74.009411, 40.741154],
-              [-74.009585, 40.741175],
-              [-74.009577, 40.741211],
-              [-74.012162, 40.741492],
-              [-74.012119, 40.741749],
-              [-74.012117, 40.741759],
-              [-74.009527, 40.741491],
-              [-74.009519, 40.74149],
-              [-74.009448, 40.741892],
-              [-74.009447, 40.741892],
-              [-74.009278, 40.741879],
-              [-74.009161, 40.742509],
-              [-74.00881, 40.742378],
-            ],
-          ],
-        ],
-      },
-    },
-  ],
-};
